% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{clusterSampleHeatmap2}
\alias{clusterSampleHeatmap2}
\title{Compute and plot column correlation matrix. Normally used
to do sample corealtion of gene expression or OTU abundance matrix.}
\usage{
clusterSampleHeatmap2(
  mat,
  method = "pearson",
  digits = 4,
  cor_file = NULL,
  saveplot = NULL,
  ...
)
}
\arguments{
\item{mat}{A dataframe.}

\item{method}{Type of correlation coefficient given to \code{\link{cor}}.
Default "pearson".}

\item{digits}{Number of decimial digits (given to \code{\link{round}}) to keep (default 4).}

\item{cor_file}{Save ordered correlation matrix to given file name.}

\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{...}{Additional parameters given to plot output (\code{\link{pdf}}, \code{\link{png}},...) like "width", "height", .etc.}
}
\value{
Nothing
}
\description{
Compute and plot column correlation matrix. Normally used
to do sample corealtion of gene expression or OTU abundance matrix.
}
\examples{

df = generateAbundanceDF()
clusterSampleHeatmap2(df)

}
