% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{clusterSamplePheatmap}
\alias{clusterSamplePheatmap}
\title{Compute and plot column correlation matrix. Normally used
to do sample corealtion of gene expression or OTU abundance matrix.}
\usage{
clusterSamplePheatmap(
  mat,
  method = "pearson",
  digits = 4,
  sampleAnno = NA,
  cellsize = NULL,
  cor_file = NULL,
  saveplot = NA,
  ...
)
}
\arguments{
\item{mat}{A dataframe.}

\item{method}{Type of correlation coefficient given to \code{\link{cor}}.
Default "pearson".}

\item{digits}{Number of decimial digits (given to \code{\link{round}}) to keep (default 4).}

\item{sampleAnno}{Add sample attribute to plot. Accept a dataframe with
rownames as "\code{colnames(mat)}".}

\item{cor_file}{Save ordered correlation matrix to given file name.}

\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{...}{Additional parameters given to \code{\link[pheatmap]{pheatmap}} like "width", "height", .etc.}

\item{Size}{of cells. Default autodetect. This is used to get square plot.}
}
\value{
Nothing
}
\description{
Compute and plot column correlation matrix. Normally used
to do sample corealtion of gene expression or OTU abundance matrix.
}
\examples{

df = generateAbundanceDF()
clusterSamplePheatmap(df)

}
