% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{clusterSampleUpperTriPlot}
\alias{clusterSampleUpperTriPlot}
\title{Compute and plot column correlation matrix. Normally used
to do sample corealtion of gene expression or OTU abundance matrix.}
\usage{
clusterSampleUpperTriPlot(
  mat,
  method = "pearson",
  digits = 4,
  cor_file = NULL,
  saveplot = NULL,
  width = 13.5,
  height = 15,
  ...
)
}
\arguments{
\item{mat}{A dataframe.}

\item{method}{Type of correlation coefficient given to \code{\link{cor}}.
Default "pearson".}

\item{digits}{Number of decimial digits (given to \code{\link{round}}) to keep (default 4).}

\item{cor_file}{Save ordered correlation matrix to given file name.}

\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{width}{Picture width in "cm".}

\item{height}{Picture height in "cm".}

\item{...}{Additional parameters given to \code{\link[ggplot2]{ggsave}}.}
}
\value{
A ggplot2 object
}
\description{
Compute and plot column correlation matrix. Normally used
to do sample corealtion of gene expression or OTU abundance matrix.
}
\examples{

df = generateAbundanceDF()
clusterSampleUpperTriPlot(df)

}
