% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{dataFilter}
\alias{dataFilter}
\title{Filter low variance genes by given minimal \code{\link{mad}} value or keep top
number/percent genes with bigger variances.}
\usage{
dataFilter(
  datExpr,
  minimal_mad = NULL,
  top_mad_n = 0.75,
  rmVarZero = T,
  noLessThan = NULL
)
}
\arguments{
\item{datExpr}{Normal gene expression matrix (gene x sample).}

\item{minimal_mad}{Minimal allowed mad value.}

\item{top_mad_n}{An integer larger than 1 will be used to get top x genes (like top 5000).
A float number less than 1 will be used to get top x fraction genes (like top 0.7 of
all genes).}

\item{rmVarZero}{Default TRUE. Remove genes with variance as 0. Normally for PCA or
correlation analysis.}

\item{noLessThan}{Specify the lowest number of genes to be kept. Default \code{NULL} meaining no lower limit.}
}
\value{
A dataframe.
}
\description{
Filter low variance genes by given minimal \code{\link{mad}} value or keep top
number/percent genes with bigger variances.
}
\examples{

df = generateAbundanceDF(nSample=30, nGrp=3)
dataFilter(df)

}
