% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{dataFilter2}
\alias{dataFilter2}
\title{Filter low variance genes by given minimal \code{\link{mad}} or \code{\link{var}} or
other statistical value or keep top
number/percent genes with bigger variances.}
\usage{
dataFilter2(
  datExpr,
  minimal_threshold = NULL,
  top_n = 1,
  rmVarZero = F,
  noLessThan = NULL,
  statistical_value_type = mad,
  keep_filtered_as_others = F,
  fillna = NULL
)
}
\arguments{
\item{datExpr}{Normal gene expression matrix (gene x sample).}

\item{minimal_threshold}{Minimal allowed statistical value.}

\item{top_n}{An integer larger than 1 will be used to get top x genes (like top 5000).
A float number less than 1 will be used to get top x fraction genes (like top 0.7 of
all genes).}

\item{rmVarZero}{Default TRUE. Remove genes with variance as 0. Normally for PCA or
correlation analysis.}

\item{noLessThan}{Specify the lowest number of genes to be kept. Default \code{NULL} meaning no lower limit.}

\item{statistical_value_type}{Specify the way for statistical computation. Default mad, accept mean, var, sum, median.}

\item{keep_filtered_as_others}{Get sums of all filtered items as an new item - Others. Default FALSE.}

\item{fillna}{Fill NA with specified value.}
}
\value{
A dataframe.
}
\description{
Filter low variance genes by given minimal \code{\link{mad}} or \code{\link{var}} or
other statistical value or keep top
number/percent genes with bigger variances.
}
\examples{

df = generateAbundanceDF(nSample=30, nGrp=3)
dataFilter2(df)

}
