% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_pheatmap.R
\name{draw_colnames_custom}
\alias{draw_colnames_custom}
\title{Pheatmap function only for inner usages}
\usage{
draw_colnames_custom(coln, gaps, xtics_angle = 0, ...)
}
\arguments{
\item{coln}{Nothing}

\item{gaps}{Nothing}

\item{xtics_angle}{Nothing}

\item{...}{Nothing}
}
\value{
A grob
}
\description{
Pheatmap function only for inner usages
}
\examples{

#Ignore

}
