% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrichCustomizedPathway}
\alias{enrichCustomizedPathway}
\title{Enrichment of customrized pathway or annotation for any organism.}
\usage{
enrichCustomizedPathway(
  de_file,
  anno_file,
  output_prefix = NULL,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2
)
}
\arguments{
\item{de_file}{Two columns file with first column containing DE genes and second column with group names.}

\item{anno_file}{Two columns file with first column of annotations or pathways and second columns of genes within this annotation or pathway.}

\item{output_prefix}{Output prefix.}

\item{pvalueCutoff}{Default 0.05}

\item{qvalueCutoff}{Default 0.2}

\item{setReadable}{Transfer gene ids to gene symbol. Default True.}

\item{typeL}{A vector with default ad c("BP", "MF", "CC").}
}
\description{
Enrichment of customrized pathway or annotation for any organism.
}
\examples{
enrichCustomizedPathway(de_file, output_prefix=output_prefix)
}
