% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrichGO_model}
\alias{enrichGO_model}
\title{GO enrichment for model organism.}
\usage{
enrichGO_model(
  de_file,
  org_db = "org.Hs.eg.db",
  output_prefix = NULL,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2,
  setReadable = TRUE,
  typeL = c("BP", "MF", "CC")
)
}
\arguments{
\item{de_file}{Two columns file with first column containing DE genes and second column with group names.}

\item{org_db}{R annotation package like org.Hs.eg.db.}

\item{output_prefix}{Output prefix.}

\item{pvalueCutoff}{Default 0.05}

\item{qvalueCutoff}{Default 0.2}

\item{setReadable}{Transfer gene ids to gene symbol. Default True.}

\item{typeL}{A vector with default ad c("BP", "MF", "CC").}
}
\description{
GO enrichment for model organism.
}
\examples{
enrichGO_model(de_file, output_prefix=output_prefix)
}
