% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrichKEGG_model}
\alias{enrichKEGG_model}
\title{KEGG enrichment for model organism.}
\usage{
enrichKEGG_model(
  de_file,
  org_db = "org.Hs.eg.db",
  output_prefix = NULL,
  organism = "human",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2,
  setReadable = TRUE
)
}
\arguments{
\item{org_db}{R annotation package like org.Hs.eg.db.}

\item{output_prefix}{Output prefix.}

\item{organism}{KEGG supported organisms like human, mouse.}

\item{pvalueCutoff}{Default 0.05}

\item{qvalueCutoff}{Default 0.2}

\item{setReadable}{Transfer gene ids to gene symbol. Default True.}

\item{Two}{columns file with first column containing DE genes and second column with group names.}
}
\value{
A dataframe containing top 10 enriched terms.
}
\description{
KEGG enrichment for model organism.
}
\examples{
enrichKEGG_model(de_file, output_prefix=output_prefix)
}
