% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_pheatmap.R
\name{find_coordinates}
\alias{find_coordinates}
\title{Pheatmap function only for inner usages}
\usage{
find_coordinates(n, gaps, m = 1:n)
}
\arguments{
\item{n}{Nothing}

\item{gaps}{Nothing}

\item{m}{Nothing}
}
\value{
A list
}
\description{
Pheatmap function only for inner usages
}
\examples{

#Ignore

}
