% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowerplot.R
\name{flower_plot_inner}
\alias{flower_plot_inner}
\title{Flower plot could be treated as one kind of venn diagram but only showing common items
like OTUs or genes among all groups and total items (or total items excluding common
items) for each group.}
\usage{
flower_plot_inner(
  sample,
  total_num,
  core_num,
  start = 90,
  a = 0.5,
  b = 2,
  r = 1,
  group_color = rgb(135, 206, 235, 150, max = 255),
  group_color_alpha = 0.6,
  common_color_alpha = 0.6,
  label = "core",
  common_color = "white",
  ...
)
}
\arguments{
\item{sample}{A vector of sample names.

Like

\if{html}{\out{<div class="sourceCode">}}\preformatted{c("Grp1", "Grp2", "Grp3")
}\if{html}{\out{</div>}}}

\item{total_num}{Number of total or specififc items for each group.

like

\if{html}{\out{<div class="sourceCode">}}\preformatted{c(20, 30, 40)
}\if{html}{\out{</div>}}}

\item{core_num}{Number of items common to all groups.}

\item{start}{Start position of first ellipse. Default 90 represents starting from 0 clock.}

\item{r}{Set the size of the center circle.}

\item{group_color}{Set the color of the petal ellipse (each group), with input format，like：c('#6181BD4E','#F348004E','#64A10E4E'...) or
supply a RColorBrewer color set like "Set1", "Set2", "Set3", "YlOrRd"
(check http://www.sthda.com/english/wiki/colors-in-r for more).}

\item{label}{The name of the center circle.}

\item{common_color}{The color of the center circle. Default "white".}
}
\value{
A pdf image.
}
\description{
Modified from http://blog.sciencenet.cn/blog-3406804-1159241.html
}
\details{
This function is not planned for public usages.
}
\examples{
flower_plot_inner(sample = sample_id, total_num = total_num, core_num = core_num)

}
