% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_shapes}
\alias{generate_shapes}
\title{Generate shape symbols for large number of groups for ggplot2}
\usage{
generate_shapes(data, shape_variable)
}
\arguments{
\item{data}{A data frame}

\item{shape_variable}{The variable treated as shape groups}
}
\value{
A vector contains all group symbols
}
\description{
Generate shape symbols for large number of groups for ggplot2
}
\examples{

# Not run

}
