% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_lower_tri}
\alias{get_lower_tri}
\title{Get lower triangle of the correlation matrix (from web)}
\usage{
get_lower_tri(cormat)
}
\arguments{
\item{cormat}{A data frame}
}
\value{
A data frame
}
\description{
Get lower triangle of the correlation matrix (from web)
}
\examples{

df = generateAbundanceDF()
df_cor = Matrix2colCorrelation(df)
get_lower_tri(df_cor)

}
