% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_upper_tri}
\alias{get_upper_tri}
\title{Get upper triangle of the correlation matrix (from web)}
\usage{
get_upper_tri(cormat)
}
\arguments{
\item{cormat}{A data frame}
}
\value{
A data fram
}
\description{
Get upper triangle of the correlation matrix (from web)
}
\examples{

df = generateAbundanceDF()
df_cor = Matrix2colCorrelation(df)
get_upper_tri(df_cor)

}
