% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match_two_df}
\alias{match_two_df}
\title{Extract same rows of two data.frame with same order.}
\usage{
match_two_df(df1, df2, way = "row-row")
}
\arguments{
\item{df1}{Dataframe 1}

\item{df2}{Dataframe 2}

\item{way}{\code{row-row (default)}: Extract same rows of two data.frame with same order.
\code{col-row}: Extarct same columns in df1 with rows in df2 and with same order}
}
\value{
A list in format like list(df1=df1, df2=df2)
}
\description{
Extract same rows of two data.frame with same order.
}
