% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Matrix2colCorrelation}
\alias{Matrix2colCorrelation}
\title{Get ordered column correlation matrix from input dataframe. Normally used
to do sample corealtion of gene expression or OTU abundance matrix.}
\usage{
Matrix2colCorrelation(mat, method = "pearson", digits = 4, cor_file = NULL)
}
\arguments{
\item{mat}{A dataframe.}

\item{method}{Type of correlation coefficient given to \code{\link{cor}}.
Default "pearson".}

\item{digits}{Number of decimial digits (given to \code{\link{round}}) to keep (default 4).}

\item{cor_file}{Save ordered correlation matrix to given file name.}
}
\value{
A list containing ordered column correlation matrix and hcluster result.
}
\description{
Get ordered column correlation matrix from input dataframe. Normally used
to do sample corealtion of gene expression or OTU abundance matrix.
}
\examples{

df = generateAbundanceDF()
Matrix2colCorrelation(df)

}
