% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{merge_data_with_auto_matched_column}
\alias{merge_data_with_auto_matched_column}
\title{Check matched columns between two data frames and try to merge them.}
\usage{
merge_data_with_auto_matched_column(df1, df2, ...)
}
\arguments{
\item{df1}{Dataframe 1}

\item{df2}{Dataframe 2}

\item{...}{Extra parameters given to \code{\link[base]{merge}}.}
}
\value{
merged dataframe
}
\description{
Check matched columns between two data frames and try to merge them.
}
\examples{

vec1 <- data.frame(col1=c('a','a','b','d'), a=c(1,2,3,6))
vec2 <- data.frame(col2=c('d','d','a','b'), b=c(1,2,4,5),a=c(1,2,3,4))
merge_data_with_auto_matched_column(vec1, vec2)

}
