% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mixedToFloat}
\alias{mixedToFloat}
\title{Transfer numeric string to numeric.}
\usage{
mixedToFloat(x)
}
\arguments{
\item{x}{A string or a vector}
}
\value{
A number or a numeric vector
}
\description{
Transfer numeric string to numeric.
}
\examples{

mixedToFloat(3)

mixedToFloat("-1/3")

mixedToFloat(c("1","0.2","1/3","-1"))

}
