% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptome.R
\name{normalizedExpr2DistribBoxplot}
\alias{normalizedExpr2DistribBoxplot}
\title{Plot distribution of normalzied expression to check the normalization effect of DESeq2.}
\usage{
normalizedExpr2DistribBoxplot(normexpr, saveplot = NULL, ...)
}
\arguments{
\item{normexpr}{A list returned by \code{\link{deseq2normalizedExpr}}.}

\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{...}{Additional parameters given to \code{\link{ggsave}}.}
}
\value{
A ggplot2 object
}
\description{
Plot distribution of normalzied expression to check the normalization effect of DESeq2.
}
\examples{


normexpr <- deseq2normalizedExpr(dds)
normalizedExpr2DistribBoxplo(normexpr)


}
