% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{numCheck}
\alias{numCheck}
\title{Check if given string or vector is all numeric}
\usage{
numCheck(x)
}
\arguments{
\item{x}{A string or a vector}
}
\value{
TRUE or FALSE
}
\description{
Check if given string or vector is all numeric
}
\examples{

numCheck(3)

numCheck("-1/3")

numCheck(c("1","0.2","1/3","-1"))

}
