% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_run}
\alias{pca_run}
\title{Title}
\usage{
pca_run(rlogMat, sample, ehbio_output_prefix)
}
\arguments{
\item{rlogMat}{xx}

\item{sample}{xx}

\item{ehbio_output_prefix}{xx}
}
\value{
xx
}
\description{
Title
}
\examples{

pca_run(data)

}
