% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rankPlot}
\alias{rankPlot}
\title{Rankplot for given column.}
\usage{
rankPlot(
  data,
  order_col = "log2FoldChange",
  midpoint = 0,
  saveplot = NULL,
  label = NULL,
  alpha = 0.5,
  colorvector = c("green", "yellow", "red"),
  width = 13.5,
  height = 15,
  ...
)
}
\arguments{
\item{data}{A dataframe with effective row names.}

\item{order_col}{Specify which column would be used for plot. Default "log2FoldChange".}

\item{midpoint}{Specify the midpoint for color show. Default 0.}

\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{label}{Label points. Accept a number to get top x points to label (both directions).
Or a vector matched with rownames of dataframe to label specified points.
Or a named vector to label specified points with new names.}

\item{alpha}{Transparency value.}

\item{colorvector}{A vector with length 2 to speicfy low and high colors.
Or a vector with length 3 to specify low, middle and high colors.
Default \code{c("green","yellow","red")}.}

\item{width}{Picture width in "cm".}

\item{height}{Picture height in "cm".}

\item{...}{Additional parameters given to \code{\link[pheatmap]{pheatmap}}
like "width", "height", .etc.}
}
\value{
A ggplot2 object
}
\description{
Rankplot for given column.
}
\examples{

a <- data.frame(log2FoldChange=rnorm(1000), row.names=paste0("ImageGP",1:1000))

## Raw plot

rankPlot(a)

## Label top 10

rankPlot(a, label=10)

## Label specified points

rankPlot(a, label=c("ImageGP1","ImageGP2","ImageGP10"))

## Label specified points with new names

b <- c("A","B","C")
names(b) <- c("ImageGP1","ImageGP2","ImageGP10")
rankPlot(a, label=b)

}
