% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptome.R
\name{salmon2deseq}
\alias{salmon2deseq}
\title{Iniitialize a DESeq2 object from salmon output.}
\usage{
salmon2deseq(
  salmon_file_list,
  sampleFile,
  design,
  covariate = NULL,
  tx2gene = NULL,
  filter = NULL,
  rundeseq = T
)
}
\arguments{
\item{salmon_file_list}{A two-column file with the first column as sample names and
second column containing the path of quant.sf generated by salmon. Header line
is required but column names do not matter.

Below is an example (pay attention to the \strong{path} of quant.sf)\preformatted{Samp    path_quant.sf
untrt_N61311    untrt_N61311/untrt_N61311.salmon.count/quant.sf
untrt_N052611    untrt_N052611/untrt_N052611.salmon.count/quant.sf
untrt_N080611    untrt_N080611/untrt_N080611.salmon.count/quant.sf
untrt_N061011    untrt_N061011/untrt_N061011.salmon.count/quant.sf
trt_N61311    trt_N61311/trt_N61311.salmon.count/quant.sf
trt_N052611    trt_N052611/trt_N052611.salmon.count/quant.sf
trt_N080611    trt_N080611/trt_N080611.salmon.count/quant.sf
trt_N061011    trt_N061011/trt_N061011.salmon.count/quant.sf
}}

\item{sampleFile}{A file containing at least two columns. The first column is sample name just
like the first column of \code{salmon_file_list}. Other columns are sample attributes.
Normally one of sample attributes should contain the group information each sample belongs to.

One simple example (conditions represent group information)\preformatted{Samp    conditions
untrt_N61311    untrt
untrt_N052611    untrt
untrt_N080611    untrt
untrt_N061011    untrt
trt_N61311    trt
trt_N052611    trt
trt_N080611    trt
trt_N061011    trt
}

Another example (3rd column meaning samples from two batches)\preformatted{Samp    conditions  batch
untrt_N61311    untrt A
untrt_N052611    untrt A
untrt_N080611    untrt B
untrt_N061011    untrt B
trt_N61311    trt A
trt_N052611    trt A
trt_N080611    trt B
trt_N061011    trt B
}}

\item{design}{A column name from "sampleFile" like "conditions" in example.
This will be used as group variable for DE tests. Currently only simple
design is allowed. If one wants to model multiple variables, construct
one representation of super variable as indicated in
\url{https://support.bioconductor.org/p/67600/#67612} may be useful.}

\item{covariate}{Names of columns containing informations maybe covariates
like batch effects or other sample info. Multiple covariates should be
supplied as a vector.}

\item{tx2gene}{Optional and only used if one want to get gene expression
instead of transcript expression.
A two-column file with the first column as transcript names and
second column as gene names. Header line is required but column names do not matter.

Below is an example of file contents.\preformatted{txname    gene
ENST00000456328    ENSG00000223972
ENST00000450305    ENSG00000223972
ENST00000488147    ENSG00000227232
ENST00000619216    ENSG00000278267
ENST00000473358    ENSG00000243485
}}

\item{filter}{Filter genes with low read counts. Default genes with total
reads count lower than half of number of samples will be filtered out.
One can give any number here. Normally default is OK. The DESeq2 will ao
auto filter too. Check \url{https://www.bioconductor.org/packages/devel/bioc/vignettes/DESeq2/inst/doc/DESeq2.html}.}

\item{rundeseq}{Default \code{TRUE}. The function will perfrom deseq analysis
using \code{\link[DESeq2]{DESeq}} and return analyzed DESeqDataSet object.
If \code{FALSE}, just return a DESeqDataSet object and one can
run \code{\link[DESeq2]{DESeq}}on it with more customed parameters.}
}
\value{
A DESeqDataSet object.
}
\description{
Iniitialize a DESeq2 object from salmon output.
}
\examples{

dds <- salmon2deseq(salmon_file_list, sampleFile, "conditions")
dds <- salmon2deseq(salmon_file_list, tx2gene=tx2gene,
                    sampleFile, "conditions")


}
