% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp.is.null}
\alias{sp.is.null}
\title{Check Null Object}
\usage{
sp.is.null(x)
}
\arguments{
\item{x}{\code{NULL} object or \code{'null'} string}
}
\value{
True when x is \code{NULL} or \code{"NULL"} (case insensitive for character type)
}
\description{
Check Null Object
}
\examples{

sp.is.null('NULL')

}
