% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_barplot.R
\name{sp_barplot}
\alias{sp_barplot}
\title{Generating bar plot}
\usage{
sp_barplot(
  data,
  color_variable = NULL,
  yvariable = NULL,
  xvariable = NULL,
  melted = TRUE,
  title = NULL,
  x_label = NULL,
  y_label = NULL,
  color_variable_order = NULL,
  xvariable_order = NULL,
  y_add = 0,
  group_variable = NULL,
  add_bar_link = FALSE,
  add_point = F,
  yaxis_scale_mode = NULL,
  facet_variable = NULL,
  stat = "identity",
  bar_mode = "stack",
  facet_variable_order = NULL,
  facet_nrow = NULL,
  facet_ncol = NULL,
  error_bar_variable = NULL,
  base_font_size = 11,
  legend.position = "right",
  xtics = TRUE,
  xtics_angle = 0,
  statistics = FALSE,
  ytics = TRUE,
  manual_color_vector = "Set2",
  facet_scales = "fixed",
  extra_ggplot2_cmd = NULL,
  coordinate_flip = FALSE,
  add_text = FALSE,
  font_path = NULL,
  debug = FALSE,
  filename = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame or data file (with header line, the first column will
not be treated as row names, tab separated).}

\item{color_variable}{Name for specifying bars colors (one of column names, should be specified
when inputting long format matrix).}

\item{yvariable}{Name for value column (one of column names, should be specified
when inputting long format matrix).}

\item{xvariable}{Name for x-axis variable (one of column names, should be specified
when inputting long format matrix).}

\item{melted}{\code{TRUE} for dealing with long format matrix, the program will skip melt preprocess. If input is wide format matrix, this parameter should be set to \code{FALSE}.}

\item{title}{Title of picture.}

\item{x_label}{Xlab label.}

\item{y_label}{Ylab label.}

\item{color_variable_order}{Set orders of color variable (this can also used to extract specific rows).}

\item{xvariable_order}{Levels for x-axis variable, suitable when x-axis is not used as a number.}

\item{y_add}{A number to add if log scale is used.
Default 0 meaning the minimum non-zero value would be used.}

\item{group_variable}{Specify group info for for computing means and SDs.}

\item{add_point}{Set to TRUE to add each point (specially used when displaying mean values)}

\item{yaxis_scale_mode}{Give the following \code{scale_y_log10()},
\code{coord_trans(y="log10")}, or other legal command for ggplot2 or
simply \code{log2} to set the scale way.}

\item{facet_variable}{Wrap plots by given column (one of column names should be specified).
This is used to put multiple plot in one picture.}

\item{stat}{The ways to show the height of bars.
The height of bars represent the numerical values in each group by default (normally in \code{yvariable} column of melted data).
One can also give \code{count} to let the program count the number of
items in each group (Normally the \code{color_variable} column is used to group
'xvariable' column after melt).
Or one can give \code{weight} which will sum values of each group.
Default \code{identity}, accept \code{count} when categorical data are given.}

\item{bar_mode}{The ways to place multiple bars for one group.
Multiple bars in same place will be stacked together by default.
Giving \code{fill} to get stacked percent bar-plot.
Giving \code{dodge} to arrange multiple bars side-by-side.
Default \code{stack}, accept \code{dodge}, \code{fill}.}

\item{facet_variable_order}{The levels of wrapping to set the order of each group.}

\item{facet_nrow}{The number of rows one want when \code{facet} is used. Default NULL.}

\item{facet_ncol}{The number of columns one want when \code{facet} is used. Default NULL.}

\item{error_bar_variable}{Error-bar column (one of column names). Specify the column containing error bars.}

\item{base_font_size}{Font-size. Default 11.}

\item{legend.position}{Position of legend, accept top, bottom, left, right, none or c(0.8,0.8).}

\item{xtics}{Display xtics. Default TRUE.}

\item{xtics_angle}{Rotation angle for a-axis. Default 0.}

\item{ytics}{Display ytics. Default FALSE.}

\item{manual_color_vector}{Manually set colors for each geom.
Default NULL, meaning using ggplot2 default.
Colors like c('red', 'blue', '#6181BD') (number of colors not matter) or
a RColorBrewer color set like  "BrBG"     "PiYG"     "PRGn"     "PuOr"
"RdBu"     "RdGy"     "RdYlBu"   "RdYlGn"  "Spectral" "Accent"
"Dark2"    "Paired"   "Pastel1"  "Pastel2"  "Set1"
"Set2"    "Set3"     "Blues"    "BuGn"     "BuPu"
"GnBu"     "Greens"   "Greys"    "Oranges" "OrRd"     "PuBu"
"PuBuGn"   "PuRd"     "Purples"  "RdPu"     "Reds"
"YlGn"    "YlGnBu"   "YlOrBr"   "YlOrRd"
(check http://www.sthda.com/english/wiki/colors-in-r for more).}

\item{facet_scales}{Paramter for scales for facet. Default \code{fixed} meaning each inner graph
use same scale (x,y range), \code{free} (variable x, y ranges for each sub-plot),
\code{free_x} (variable x ranges for each sub-plot), \code{free_y} (variable y ranges for each sub-plot).}

\item{extra_ggplot2_cmd}{Other legal R codes for ggplot2 will be given here.}

\item{coordinate_flip}{Flip cartesian coordinates so that horizontal becomes vertical, and vertical, horizontal. This is primarily useful for converting geoms and statistics which display y conditional on x, to x conditional on y.}

\item{add_text}{Add text to bar. Default FALSE.}

\item{font_path}{Specify font type. Give a path for one font type file
like '/etc/fonts/Arial.ttf'
or 'HeiArial.ttc'(if in current directory), Default system default.}

\item{filename}{Output picture to given file.}

\item{...}{}
}
\value{
A ggplot2 object
}
\description{
Generating bar plot
}
\examples{

bar_test_data <- data.frame(ID = letters[1:4], Gene = letters[c(8,8,9,9,10,10,11,11)], Expr = runif(16))
sp_barplot(data = bar_test_data, xvariable = "ID", yvariable = "Expr", color_variable = "Gene")

## Not run:
bar_data = "bar.data"

sp_barplot(data = bar_data, xvariable = "ID", yvariable = "Expr", color_variable = "Gene")
## End(Not run)

}
