% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_corrplot.R
\name{sp_corrplot}
\alias{sp_corrplot}
\title{A visualization of a correlation matrix.}
\usage{
sp_corrplot(
  data,
  method = "circle",
  type = "full",
  bg = "white",
  title = NULL,
  is.corr = FALSE,
  diag = TRUE,
  addCoef.col = NULL,
  addCoefasPercent = FALSE,
  order = "original",
  hclust.method = NULL,
  tl.pos = "n",
  tl.cex = 1,
  tl.col = "red",
  tl.srt = 90,
  cl.pos = NULL,
  cl.lim = NULL,
  cl.align.text = "c",
  saveplot = NULL,
  mar = c(0, 0, 0, 0),
  ...
)
}
\arguments{
\item{data}{Matrix or data file (with header line, the first column will be treated as row names, tab separated).}

\item{method}{The visualization method of correlation matrix to be used. Currently, it supports seven methods, named "circle" (default), "square", "ellipse", "number", "pie", "shade" and "color".}

\item{type}{Type, "full" (default), "upper" or "lower", display full matrix, lower triangular or upper triangular matrix.}

\item{bg}{The background color.}

\item{title}{Title of the graph.}

\item{is.corr}{Logical, whether the input matrix is a correlation matrix or not. We can visualize the non-correlation matrix by setting is.corr = FALSE.}

\item{diag}{Logical, whether display the correlation coefficients on the principal diagonal.}

\item{addCoef.col}{Color of coefficients added on the graph. If NULL (default), add no coefficients.}

\item{addCoefasPercent}{Logic, whether translate coefficients into percentage style for spacesaving.}

\item{order}{The ordering method of the correlation matrix. "original" for original order (default). "AOE" for the angular order of the eigenvectors. "FPC" for the first principal component order. "hclust" for the hierarchical clustering order. "alphabet" for alphabetical order.}

\item{hclust.method}{The agglomeration method to be used when order is hclust. This should be one of "ward", "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median" or "centroid".}

\item{tl.cex}{Numeric, for the size of text label (variable names).}

\item{tl.col}{The color of text label.}

\item{tl.srt}{Numeric, for text label string rotation in degrees.}

\item{cl.pos}{Character or logical, position of color labels; If character, it must be one of "r" (default if type=="upper" or "full"), "b" (default if type=="lower") or "n", "n" means don't draw colorlabel.}

\item{cl.lim}{The limits (x1, x2) in the colorlabel.}

\item{cl.align.text}{"l", "c" (default) or "r", for number-label in colorlabel, "l" means left, "c" means center, and "r" means right.}

\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{...}{}
}
\value{
a grid object
}
\description{
A visualization of a correlation matrix.
}
\examples{
data<-matrix(rnorm(100),nrow=20)
rownames(data)<- paste0("corrtest", 1:20)
colnames(data)<- paste0("zhou", 1:5)
sp_corrplot(data, cl.align.text = "l")
sp_corrplot(data,method="pie",type="lower",title="corrplot", cl.align.text = "l", mar=c(0,0,1,0))
sp_corrplot(data,method="ellipse",type="upper", cl.align.text = "l", tl.pos="td")
}
