% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_current_time}
\alias{sp_current_time}
\title{Get current time in strign format}
\usage{
sp_current_time(delim_left = "[", delim_right = "]")
}
\arguments{
\item{delim_left}{Default \code{[}.}

\item{delim_right}{Default \verb{]}.}
}
\value{
A string
}
\description{
Get current time in strign format
}
\examples{

sp_current_time()

}
