% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_dendrogram.R
\name{sp_dendextend}
\alias{sp_dendextend}
\title{Hierarchical cluster diagram}
\usage{
sp_dendextend(
  data,
  group_variable = NULL,
  branch_order = NULL,
  method = "complete",
  k = k,
  labels_size = 0.5,
  shape = "tree",
  pic_title = NULL,
  pic_flip = TRUE,
  node_size = 0.007,
  legend_site = "topleft",
  saveplot = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{group_variable}{Specifies a column as group.}

\item{branch_order}{Specify branch order.}

\item{method}{The agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{k}{Number of groups.}

\item{labels_size}{Labels size.}

\item{shape}{Tree or circles.}

\item{pic_title}{Title of the graph.}

\item{pic_flip}{TRUE for horizontal, FALSE for vertical. Default TRUE.}

\item{node_size}{Node size.}

\item{legend_site}{Legend site. Optional, "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center".}

\item{...}{}
}
\description{
Hierarchical cluster diagram
}
\examples{
data<-matrix(rnorm(200),nrow = 50)
rownames(data)<- paste0("dendextend", 1:50)
colnames(data)<- paste0("zhou", 1:4)
sp_dendextend(data = data,k = 3,labels_size = 0.3)


data<- data.frame(ID = letters[1:5], apple = runif(50), banana = runif(50), watermelon = runif(50))
sp_dendextend(data = data, k = 5, method = "single", shape = "circle")

}
