% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{sp_diff_test}
\alias{sp_diff_test}
\title{Perform statistical test using given methods.}
\usage{
sp_diff_test(
  data,
  stat_value_variable,
  stat_group_variable,
  statistical_method = "aov",
  add_y = TRUE,
  statistical_threshold_for_letters = 0.05
)
}
\arguments{
\item{data}{A data matrix}

\item{stat_value_variable}{The column represents the statistical value information.}

\item{stat_group_variable}{The column represents the statistical group information.}

\item{statistical_method}{Statistical method. For two groups, default <t.test>. For more than two groups, default \if{html}{\out{<aov>}}.}

\item{add_y}{Add positions for each statistical label.}

\item{statistical_threshold_for_letters}{Threshold for treating as significance, default 0.05.}
}
\value{
A list. list(data=data, a data frame with statistical information, Tukey_HSD=Tukey_HSD)
}
\description{
Perform statistical test using given methods.
}
\examples{
NULL


}
