% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_enrichment.R
\name{sp_enrichment}
\alias{sp_enrichment}
\title{Generating enrichment plot}
\usage{
sp_enrichment(
  data,
  xvariable,
  yvariable,
  size_variable = NULL,
  color_variable = NULL,
  xvariable_order = NULL,
  yvariable_order = NULL,
  shape_variable_order = NULL,
  title = NULL,
  x_label = NULL,
  y_label = NULL,
  yvariable_width = 60,
  manual_color_vector = c("green", "red"),
  log10_transform_variable = "no_transform",
  sqrt_transform_variable = "no_transform",
  legend.position = "right",
  xtics_angle = 0,
  shape_variable = NULL,
  coordinate_flip = FALSE,
  extra_ggplot2_cmd = NULL,
  file_name = NULL,
  ...
)
}
\arguments{
\item{data}{Data file (with header line, the first column is not the rowname, tab seperated).}

\item{xvariable}{One of column names used as the variable for horizontal axis.}

\item{yvariable}{One of column names used as the variable for vertical axis.}

\item{size_variable}{One of column names used as the variable for defining point sizes.
Normally \code{count}. Default the variable name given to \code{sqrt_transform_variable}.}

\item{color_variable}{One of column names used as the variable for defining point colors. Default the
variable name given to \code{log10_transform_variable}.}

\item{xvariable_order}{The order for horizontal axis. Only woroks if horizontal axis
variables are strings. If horizontaol axis are numbers like GeneRatio,
this will be treated as shape_variable order. Default alphabetical order,
accept a vector like c('SampA_Up','SampB_Up').}

\item{shape_variable_order}{The order for shape variable. Default alphabetical order,
accept a vector like c('SampA_Up','SampB_Up').}

\item{title}{Title of picture. Default empty.}

\item{x_label}{X-axis title of picture. Default empty.}

\item{y_label}{Y-axis title of picture. Default empty.}

\item{yvariable_width}{Default 60. Normally yvariable would be terms. This is used to set
the max-allowed term length. Terms with length larger than given  value will be wrap
to new line.}

\item{manual_color_vector}{Manually set colors for each geom.
Default NULL, meaning using ggplot2 default.
Colors like c('red', 'blue', '#6181BD') (number of colors not matter) or
a RColorBrewer color set like  "BrBG"     "PiYG"     "PRGn"     "PuOr"
"RdBu"     "RdGy"     "RdYlBu"   "RdYlGn"  "Spectral" "Accent"
"Dark2"    "Paired"   "Pastel1"  "Pastel2"  "Set1"
"Set2"    "Set3"     "Blues"    "BuGn"     "BuPu"
"GnBu"     "Greens"   "Greys"    "Oranges" "OrRd"     "PuBu"
"PuBuGn"   "PuRd"     "Purples"  "RdPu"     "Reds"
"YlGn"    "YlGnBu"   "YlOrBr"   "YlOrRd"
(check http://www.sthda.com/english/wiki/colors-in-r for more).}

\item{log10_transform_variable}{Get log-transformed data for given variable.
Default \code{no_transform}, means no log10 transform.
Accept a variable like \code{p_value} to get \code{(-1) * log10(p_value)}.}

\item{sqrt_transform_variable}{Get square root-transformed data for given variable.
Default \code{no_transform}, means no sqrt transform.
Accept a variable like \code{count} to get \code{sqrt(count)}.}

\item{legend.position}{Position of legend, accept top, bottom, left, right, none or c(0.8,0.8).}

\item{xtics_angle}{Rotation angle for a-axis. Default 0.}

\item{shape_variable}{One of column names used as the variable for defining point shapes.}

\item{coordinate_flip}{Flip cartesian coordinates so that horizontal becomes vertical, and vertical, horizontal. This is primarily useful for converting geoms and statistics which display y conditional on x, to x conditional on y.}

\item{extra_ggplot2_cmd}{Extra ggplot2 commands (currently unsupported)}

\item{...}{Parameters given to \code{sp_ggplot_layout}}

\item{scale_size_min}{Scale size with minimum value specified}

\item{scale_size_max}{Scale size with maximum value specified}
}
\value{
A ggplot2 object
}
\description{
Generating enrichment plot
}
\examples{



## Not run:
enrichment_data <- "enrichment.data"

library(YSX)
enrichment_data <- "enrichment.data"
sp_enrichment(data = enrichment_data, xvariable = "GeneRatio", yvariable = "Description",
              log10_transform_variable = "Qvalue", sqrt_transform_variable = "Count",
              shape_variable = "SampleGroup")

enrichment_data <- "goeast.enrich.txt"
enrichment.data <- sp_readTable(enrichment_data)
head(enrichment.data)

p <- sp_enrichment(data = enrichment.data, xvariable = "log_odds_ratio",
                   yvariable = "Term", color_variable = "p",
                   log10_transform_variable="p", size_variable = "q",
                   sqrt_transform_variable = "q",
                   shape_variable = "Ontology")
## End(Not run)


}
