% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_EulerDiagrams.R
\name{sp_EulerDiagrams}
\alias{sp_EulerDiagrams}
\title{Generate Euler diagrams (proposional)}
\usage{
sp_EulerDiagrams(
  data,
  format = "items",
  intersection_variable = NULL,
  count_variable = NULL,
  item_variable = NULL,
  set_variable = NULL,
  type = c("counts"),
  shape = "circle",
  manual_color_vector = NULL,
  alpha = 1,
  legend.position = "right",
  font_quantities = 1,
  lty = 1,
  labels_font = 1,
  saveplot = NULL,
  saveppt = FALSE,
  ...
)
}
\arguments{
\item{data}{One filename or dataframe containing data in specified formats with header line.}

\item{format}{\code{items} or \code{counts} with format specified above.}

\item{intersection_variable}{Only used when \code{format=counts} to specify which column
contains different types of interactions. For example data, first column name \code{Intersection}
should be given here. Color should be specified for the appearance order of each set.}

\item{count_variable}{Only used when \code{format=counts} to specify which column
contains computed counts for different types of interactions. For example data,
first column name \code{Count} should be given here.}

\item{item_variable}{Only used when \code{format=items} to specify which column
contains all items like genes or OTUs or species. For example data, first column
name \code{Items} should be given here.}

\item{set_variable}{Only used when \code{format=items} to specify which column
contains group information of items. For example data, second column
name \code{Group} should be given here. Color should be specified for the alphabetial
order of each set.}

\item{type}{Show \code{percent} or \code{counts} in the plot. Default \code{counts}.}

\item{shape}{Use \code{circle} or \code{ellipse} in the plot. Default \code{circle}.}

\item{alpha}{Generate an alpha transparency values for return colors. 0 means fully transparent and 1 means opaque. Default 1.}

\item{legend.position}{Position of legend, accept top, bottom, left, right, none or c(0.8,0.8).}

\item{font_quantities}{Font size for numbers in Euler plot. Default \code{1}.}

\item{lty}{Line type of circle or ellipse edges from \code{1} to \code{6} represents \code{solid},
\code{dashed}, \code{dotted}, \code{dotdash}, \code{longdash} and \code{twodash} separately. Default \code{1}.}

\item{labels_font}{Font size for labels in Euler plot. Default \code{1}.}

\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{saveppt}{Output PPT format.}

\item{...}{Other parameters given to base_plot_save}
}
\value{
a grid object
}
\description{
When \code{format} is \code{items}, \code{data} examples are

\if{html}{\out{<div class="sourceCode">}}\preformatted{Items  Group
g1	Set1
g2	Set1
a1	Set3
a3	Set1
b4	Set1
g1	Set2
h1	Set4
}\if{html}{\out{</div>}}
}
\details{
When \code{format} is \code{counts}, \code{data} examples are

\if{html}{\out{<div class="sourceCode">}}\preformatted{Intersection	Count
Set1&Set2	2
Set1&Set3&Set4&Set5	1
Set3&Set5	2
Set1	1
Set2&Set4	1
Set2&Set3&Set4	1
Set2&Set5	1
Set3	1
Set5	1
Set4&Set5	1
Set4	2
}\if{html}{\out{</div>}}
}
\examples{

NULL


}
