% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_get_ggplot_limits}
\alias{sp_get_ggplot_limits}
\title{Get the x, y limits of a ggplot2 plot}
\usage{
sp_get_ggplot_limits(p)
}
\arguments{
\item{p}{A ggplot2 object}
}
\value{
A list list(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax)
}
\description{
Get the x, y limits of a ggplot2 plot
}
\examples{
## Not run:
sp_get_ggplot_limits(p)

## End(Not run)
}
