% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_ggplot_add_vline_hline}
\alias{sp_ggplot_add_vline_hline}
\title{Add hline or vline for ggplot2 object}
\usage{
sp_ggplot_add_vline_hline(
  p,
  custom_vline_x_position = NULL,
  custom_vline_anno = NULL,
  custom_vline_anno_y_pos = NULL,
  custom_hline_y_position = NULL,
  custom_hline_anno = NULL,
  custom_hline_anno_x_pos = NULL,
  linetype = "dotted",
  size = 0.5,
  ...
)
}
\arguments{
\item{p}{A ggplot2 object}

\item{custom_vline_x_position}{A vector of coordinates for vertical lines.}

\item{custom_vline_anno}{Annotation text for each vertical line.}

\item{custom_hline_y_position}{A vector of coordinates for horizontal lines.}

\item{custom_hline_anno}{Annotation text for each horizontal line.}

\item{...}{Extra parameters given to \code{geom_vline} and \code{geom_hline}}
}
\value{
A ggplot2 object
}
\description{
Add hline or vline for ggplot2 object
}
\examples{

## Not run:
sp_ggplot_add_vline_hline(p)

## End(Not run)


}
