% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_ggplot_facet}
\alias{sp_ggplot_facet}
\title{Facet ggplot2 object}
\usage{
sp_ggplot_facet(
  p,
  facet_variable = NULL,
  facet_ncol = NULL,
  facet_nrow = NULL,
  facet_scales = "fixed"
)
}
\arguments{
\item{p}{A ggplot2 object}

\item{facet_variable}{Wrap plots by given column (one of column names should be specified).
This is used to put multiple plot in one picture.}

\item{facet_ncol}{The number of columns one want when \code{facet} is used. Default NULL.}

\item{facet_nrow}{The number of rows one want when \code{facet} is used. Default NULL.}

\item{facet_scales}{Paramter for scales for facet. Default \code{fixed} meaning each inner graph
use same scale (x,y range), \code{free} (variable x, y ranges for each sub-plot),
\code{free_x} (variable x ranges for each sub-plot), \code{free_y} (variable y ranges for each sub-plot).}
}
\value{
A ggplot2 object
}
\description{
Facet ggplot2 object
}
\examples{

## Not run:
sp_ggplot_facet(p, facet_variable)

## End(Not run)

}
