% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_ggplot_layout}
\alias{sp_ggplot_layout}
\title{Change common layout of ggplot2 object}
\usage{
sp_ggplot_layout(
  p,
  xtics_angle = 0,
  legend.position = "right",
  extra_ggplot2_cmd = NULL,
  filename = NULL,
  x_label = NULL,
  y_label = NULL,
  title = NULL,
  coordinate_flip = FALSE,
  ylim = NULL,
  width = 12,
  height = 6.18,
  fontname = "",
  base_font_size = 10,
  additional_theme = NULL,
  zoom_split = FALSE,
  zoom_xlim = NULL,
  zoom_ylim = NULL,
  saveppt = FALSE,
  savehtml = FALSE,
  ...
)
}
\arguments{
\item{p}{A ggplot2 object}

\item{xtics_angle}{Rotation angle for a-axis. Default 0.}

\item{legend.position}{Position of legend, accept top, bottom, left, right, none or c(0.8,0.8).}

\item{extra_ggplot2_cmd}{Extra ggplot2 commands (currently unsupported)}

\item{filename}{Output picture to given file.}

\item{x_label}{Xlab label.}

\item{y_label}{Ylab label.}

\item{title}{Title of picture.}

\item{coordinate_flip}{Flip cartesian coordinates so that horizontal becomes vertical, and vertical, horizontal. This is primarily useful for converting geoms and statistics which display y conditional on x, to x conditional on y.}

\item{width}{Picture width (units: cm)}

\item{height}{Picture height (units: cm)}

\item{zoom_split}{If both x and y is given, should each axis zoom be shown separately as well? Defaults to FALSE.}

\item{zoom_xlim}{Specific zoom ranges for x axis.}

\item{zoom_ylim}{Specific zoom ranges for y axis.}

\item{saveppt}{Output PPT format.}

\item{savehtml}{Save the images as HTML files.}

\item{...}{Extra parameters to \code{\link[ggplot2]{ggsave}}.}
}
\value{
A ggplot2 object
}
\description{
Change common layout of ggplot2 object
}
\examples{

## Not run:
sp_ggplot_layout(p)

## End(Not run)

}
