% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_hclust.R
\name{sp_hclust}
\alias{sp_hclust}
\title{Hierarchical cluster diagram}
\usage{
sp_hclust(
  data,
  method = "average",
  thresholdZ.k = -2.5,
  saveplot = NULL,
  debug = FALSE,
  ...
)
}
\arguments{
\item{data}{A matrix file or an object.}

\item{method}{Clustering method :"ward.D", "single", "complete", "average", "mcquitty", "median", "centroid", "ward.D2"}

\item{thresholdZ.k}{Threshold for defining outliers. First compute the overall
corelation of one sample to other samples. Then do Z-score transfer for all
correlation values. The samples with corelation values less than given value
would be treated as outliers.
Default -2.5 meaning -2.5 std.}

\item{...}{}
}
\value{
A data frame.
}
\description{
Hierarchical cluster diagram
}
\examples{
x = runif(10)
y = runif(10)
data=cbind(x, y)
rownames(data) = paste("exam", 1:10)
sp_hclust(data)

}
