% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_load_font}
\alias{sp_load_font}
\title{Use showtext to load fonts}
\usage{
sp_load_font(font_path)
}
\arguments{
\item{font_path}{Specify font type. Give a path for one font type file
like '/etc/fonts/Arial.ttf'
or 'HeiArial.ttc'(if in current directory), Default system default.}
}
\value{
font_name or null
}
\description{
Use showtext to load fonts
}
\examples{

## Not run:
sp_load_font(font_path="arial.tff")

## End(Not run)

}
