% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_manhattan2.R
\name{sp_manhattan2_plot}
\alias{sp_manhattan2_plot}
\title{Generating manhattan plot}
\usage{
sp_manhattan2_plot(
  data,
  ID_var = "ID",
  FDR_var = "FDR",
  grp_var = "Phylum",
  sig_col = "level",
  status_col_level = c(),
  pvalue = 0.05,
  shape_col = NULL,
  color_var = NULL,
  alpha = 0.4,
  grp_var_order = c(),
  color_var_order = c(),
  shape_col_order = c(),
  log10_transform_fdr = TRUE,
  filename = NULL,
  point_label_var = "CTctctCT",
  title = "",
  x_label = NULL,
  y_label = "Negative log10 FDR",
  legend.position = "right",
  point_size = "ct___",
  ...
)
}
\arguments{
\item{data}{Data frame or data file (with header line, the first column will not be treated as the rowname, tab seperated).}

\item{ID_var}{Name of ID column.}

\item{FDR_var}{Name of FDR column.}

\item{grp_var}{Group variable for ordering points.}

\item{sig_col}{Significant column.Optional, in sample data significant.}

\item{status_col_level}{Changing the order of status column values. Normally, the unique values of status column would be sorted alphabetically.}

\item{pvalue}{Set the filter threshold for defining significance. Default "0.05".}

\item{shape_col}{Column for points shapes.}

\item{color_var}{Column for color points. Default the same as group variable.}

\item{alpha}{Transparent alpha value.Default 0.4, Accept a float from 0(transparent), 1(opaque).}

\item{grp_var_order}{Group variable order, like "'Rhizobiales', 'Actinomycetales'".}

\item{color_var_order}{Color variable order. Default same as \code{grp_var}.}

\item{shape_col_order}{Levels for shapes, like "'Sig','nonSig'".}

\item{log10_transform_fdr}{Get \code{-log10(FDR)} for column given to \code{FDR_var}. Default FALSE, accept TRUE.}

\item{filename}{Output picture to given file.}

\item{point_label_var}{Name of columns containing labels for points.}

\item{title}{Title of picture.}

\item{x_label}{Xlab label.Default NULL.}

\item{y_label}{Ylab label.Default "Negative log10 FDR".}

\item{legend.position}{Position of legend, accept top, bottom, left, right, none or c(0.8,0.8).}

\item{...}{}
}
\value{
A ggplot2 object
}
\description{
Generating manhattan plot
}
\examples{


## Not run:
manhattan_data = "manhattan.data"

sp_manhattan2_plot(data=manhattan_data, ID_var='ID', FDR_var='FDR', title="test1", point_size=2, point_label_var = "Labels")
## End(Not run)
}
