% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_manual_fill_ggplot2}
\alias{sp_manual_fill_ggplot2}
\title{Add manual fill-color assignment for both categorical and numerical variable}
\usage{
sp_manual_fill_ggplot2(
  p,
  data,
  color_variable,
  manual_color_vector = NULL,
  alpha = 1
)
}
\arguments{
\item{p}{A ggplot2 object}

\item{data}{Data matrix used for the ggplot2 object \code{p}}

\item{color_variable}{Name of columns for color assignment}

\item{manual_color_vector}{Manually set colors for each geom.
Default NULL, meaning using ggplot2 default.
Colors like c('red', 'blue', '#6181BD') (number of colors not matter) or
a RColorBrewer color set like  "BrBG"     "PiYG"     "PRGn"     "PuOr"
"RdBu"     "RdGy"     "RdYlBu"   "RdYlGn"  "Spectral" "Accent"
"Dark2"    "Paired"   "Pastel1"  "Pastel2"  "Set1"
"Set2"    "Set3"     "Blues"    "BuGn"     "BuPu"
"GnBu"     "Greens"   "Greys"    "Oranges" "OrRd"     "PuBu"
"PuBuGn"   "PuRd"     "Purples"  "RdPu"     "Reds"
"YlGn"    "YlGnBu"   "YlOrBr"   "YlOrRd"
(check http://www.sthda.com/english/wiki/colors-in-r for more).}

\item{alpha}{Transparency}
}
\value{
A ggplot2 object
}
\description{
Add manual fill-color assignment for both categorical and numerical variable
}
\examples{

## Not run:
p <- sp_manual_fill_ggplot2(p, data, color_variable, manual_color_vector)

## End(Not run)

}
