% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{sp_multiple_group_diff_test}
\alias{sp_multiple_group_diff_test}
\title{Perform statistical test using given methods for split data.}
\usage{
sp_multiple_group_diff_test(
  data,
  stat_value_variable,
  stat_group_variable,
  group_variable = NULL,
  ...
)
}
\arguments{
\item{data}{A data matrix}

\item{stat_value_variable}{The column represents the statistical value information.}

\item{stat_group_variable}{The column represents the statistical group information.}

\item{group_variable}{The column represents the group information.
The data would be split by this group and diff test would be performed within each group.}

\item{...}{Other parameters given to \link{sp_diff_test}.}
}
\value{
A list. list(data=data, a data frame with statistical information, Tukey_HSD=Tukey_HSD)
}
\description{
Perform statistical test using given methods for split data.
}
\examples{
NULL

}
