% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{sp_rda}
\alias{sp_rda}
\title{Generate RDA plot with species table and env table}
\usage{
sp_rda(
  otu_table,
  env_table,
  metadata,
  sample_min_reads_count = 10000,
  rare_otu_count = 5,
  otu_scale_method = "hellinger"
)
}
\arguments{
\item{otu_table}{Normalized OTU/Species abundance data frame or data file (with header line, the first column will be treated as the row names, tab separated)}

\item{env_table}{Environment factor data frame or data file (with header line, the first column will be treated as the row names, tab separated)}

\item{metadata}{Metadata file (or data.frame) with sample attributes like group information.
The first column is the same as the first row of value given to parameter \code{data}.
These attributes would be used as \code{color}, \code{size}, \code{shape} variables in the plot.
If not supplied, each sample will be treated as one group.}

\item{sample_min_reads_count}{The minimum allowed sample reads count. Default 10000 meaning samples with total reads count less than 10000 would be filtered.}

\item{rare_otu_count}{Definite rare OTU. OTU with abundance in all samples less than given value would be filtered.}

\item{otu_scale_method}{Popular (and effective) standardization methods for community ecologists
like total, max, frequency, normalize, range, rank, rrank, standardize, hellinger, chi.square, rclr,
log, clr, alr.}

\item{...}{Parameters given to \code{sp_ggplot_layout}}
}
\value{
A ggplot2 object
}
\description{
Generate RDA plot with species table and env table
}
\examples{

## Not run:
input <- "KO-OE_all.txt"
volcano_plot(input,log2fc_var='Log2FoldChange',fdr_var='Padj',status_col_var='',
             title="sd",label="Label",log10_transform_fdr=TRUE,point_size=5)
## End(Not run)
}
