% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_read_in_long_wide_matrix}
\alias{sp_read_in_long_wide_matrix}
\title{Used to read in long/wide format file or datafrmes. Wide format would be transferred to lonf fromat.}
\usage{
sp_read_in_long_wide_matrix(data, xvariable, melted)
}
\arguments{
\item{data}{Data frame or data file (with header line, the first column will
not be treated as row names for long format matrix, tab seperated).}

\item{xvariable}{Name for x-axis variable.}

\item{melted}{\code{TRUE} for dealinig with long format matrix, the program will skip melt preprocess.
Default \code{FALSE} for dealing with wide format matrix.}
}
\value{
a A long format dataframe
}
\description{
Used to read in long/wide format file or datafrmes. Wide format would be transferred to lonf fromat.
}
\examples{

## Not run:
sp_read_in_long_wide_matrix(data, xvariable, melted)

## End(Not run)

}
