% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_set_factor_order}
\alias{sp_set_factor_order}
\title{Set factor order of given variable. If \code{variable_order} is supplied, only
factors in \code{variable_order} will be kept and re-factored. Other variables
would be depleted.}
\usage{
sp_set_factor_order(data, variable, variable_order = NULL)
}
\arguments{
\item{data}{A data matrix}

\item{variable}{One column name of data matrix}

\item{variable_order}{Expected order of \code{data[[variable]]}.}
}
\value{
A data frame
}
\description{
Set factor order of given variable. If \code{variable_order} is supplied, only
factors in \code{variable_order} will be kept and re-factored. Other variables
would be depleted.
}
\examples{

data <- data.frame(A=letters[1:4], B=letters[1:4])
data
data = sp_set_factor_order(data,'A')
data$A
data = sp_set_factor_order(data,'B',c('c','d','b','a'))
data$B
data = sp_set_factor_order(data,'B',c('c','d','a'))
data$B

}
