% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_string2vector}
\alias{sp_string2vector}
\title{Transfer color string to vector}
\usage{
sp_string2vector(x, pattern = ",")
}
\arguments{
\item{x}{A string}

\item{pattern}{delimiter of sub-strings}
}
\value{
A vector
}
\description{
Transfer color string to vector
}
\examples{

sp_string2vector('red, blue,white')

}
