% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_transfer_one_column}
\alias{sp_transfer_one_column}
\title{Transfer one column of data.}
\usage{
sp_transfer_one_column(data, variable, yaxis_scale_mode = NULL, y_add = 0)
}
\arguments{
\item{data}{A data matrix}

\item{variable}{One column name of data matrix}

\item{yaxis_scale_mode}{Give the following \code{scale_y_log10()},
\code{coord_trans(y="log10")}, or other legal command for ggplot2 or
simply \code{log2} to set the scale way.}

\item{y_add}{A number to add if log scale is used.
Default 0 meaning the minimum non-zero value would be used.}
}
\value{
A data frame
}
\description{
Transfer one column of data.
}
\examples{

data <- data.frame(A=letters[1:4], B=letters[1:4])
data
data = sp_transfer_variable(data,'A', "log2")
}
