% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_vennDiagram.R
\name{sp_vennDiagram}
\alias{sp_vennDiagram}
\title{Generating venDiagram plot}
\usage{
sp_vennDiagram(
  data,
  supplyNumbers = FALSE,
  header = FALSE,
  title = NULL,
  label1 = NULL,
  label2 = NULL,
  label3 = NULL,
  label4 = NULL,
  label5 = NULL,
  color_for_circumference = "transparent",
  numVector = c(),
  labelVector = c(),
  manual_color_vector = c("dodgerblue", "goldenrod1", "darkorange1", "seagreen3",
    "orchid3"),
  alpha = 0.5,
  label_size = NULL,
  margin = NULL,
  filename = NULL,
  debug = FALSE,
  ...
)
}
\arguments{
\item{data}{Data file (the first column is the name of genes or other things one want to compare, the second column is set name of each items, tab separated).}

\item{supplyNumbers}{If you have the size of each set and the number overlaps between or among each set,  please give TRUE here.}

\item{header}{a logical value indicating whether the file contains the
    names of the variables as its first line.  If missing, the value is
    determined from the file format: \code{header} is set to \code{TRUE}
    if and only if the first row contains one fewer field than the
    number of columns.}

\item{title}{Title for the picture.}

\item{label1}{The name for label1. One string in your second column. If not supplied, the program will try to do venn plot for all sets or first 5 sets.}

\item{label2}{The name for label2.}

\item{label3}{The name for label3.}

\item{label4}{The name for label4.}

\item{label5}{The name for label5.}

\item{numVector}{List of numbers for venn plot(used when \code{supplyNumbers} is true).
For two-set venn, the format is "100, 110, 50" represents
(length_a, length_b, a_b_overlap).
For three-set venn, the format is "100, 110, 90, 50, 40, 40, 20"
represents (length_a, length_b, length_c,
a_b_overlap,  b_c_overlap, a_c_overlap, a_b_c_overlap).
For four-set venn, the format is "100, 110, 90, 50, 40, 40, 20"
represents (length_a, length_b, length_c,
a_b_overlap, a_c_overlap, a_d_overlap, b_c_overlap,
b_d_overlap, c_d_overlap, abc_overlap, abd_overlap,
acd_overlap, bcd_overlap, abcd_overlap).}

\item{labelVector}{List of label for venn plot(used when \code{supplyNumbers} is true).
Format: c('a', 'b')" for two-set and c('a', 'b', 'c') for three-set.}

\item{manual_color_vector}{Color for each area. Ussally the number of colors should
be equal to the number of labels (however the program will make them equal).
If you manually set colors for 4-way
venn diagram, the first color will be given to the
leftmost set, the second will be given to the rightmost
set, the third will be given to second leftmost and the forth
will be given to the second rightmost.
Colors like c('red', 'blue', '#6181BD') (number of colors not matter) or
a RColorBrewer color set like  "BrBG"     "PiYG"     "PRGn"     "PuOr"
"RdBu"     "RdGy"     "RdYlBu"   "RdYlGn"  "Spectral" "Accent"
"Dark2"    "Paired"   "Pastel1"  "Pastel2"  "Set1"
"Set2"    "Set3"     "Blues"    "BuGn"     "BuPu"
"GnBu"     "Greens"   "Greys"    "Oranges" "OrRd"     "PuBu"
"PuBuGn"   "PuRd"     "Purples"  "RdPu"     "Reds"
"YlGn"    "YlGnBu"   "YlOrBr"   "YlOrRd"
(check http://www.sthda.com/english/wiki/colors-in-r for more).}

\item{alpha}{Color transparency (0-1). 0: opaque; 1: transparent.}

\item{label_size}{Siez of category names. Default system default.}

\item{margin}{Number giving the amount of whitespace around the diagram in grid units. Default system default}

\item{...}{}
}
\value{
A grid object.
}
\description{
Generating venDiagram plot
}
\examples{


vennDiagram_test_data <- data.frame(elements=c("1","2","2","2","3"), sets=c("A","A","B","C","C"))
sp_vennDiagram(data = vennDiagram_test_data, label1 = "A",label2 = "B", label3 = "C")



sp_vennDiagram( supplyNumbers = TRUE,  numVector=c (120, 110, 50), labelVector=c('a','b'))



## Not run:
vennDiagram_data = "vennDiagram.data"
sp_vennDiagram(data = vennDiagram_data, label1 = "Set1",label2 = "Set2")
## End(Not run)

}
