% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_vennDiagram3.R
\name{sp_vennDiagram3}
\alias{sp_vennDiagram3}
\title{Generating venDiagram plot}
\usage{
sp_vennDiagram3(
  data,
  header = TRUE,
  item_variable = NULL,
  set_variable = NULL,
  select_set_to_show = NULL,
  doWeights = FALSE,
  type = "ellipses",
  SetLabels = TRUE,
  Faces = TRUE,
  Sets = TRUE,
  FaceText = "weight",
  saveplot = NULL,
  debug = FALSE,
  saveppt = FALSE,
  ...
)
}
\arguments{
\item{data}{Data file (the first column is the name of genes or other things one want to compare, the second column is set name of each items, tab separated).}

\item{item_variable}{Specify the column containing all items (one of column names of data).}

\item{set_variable}{Specify the column containing set names (one of column names of data).}

\item{select_set_to_show}{Specific sets to look at.}

\item{doWeights}{Whether to use weights to represent data sets}

\item{type}{Represent data sets in different shapes. Do weighted select NO if the results are wrong
For 2-set Venn diagrams: circles, squares.
For 3-set Venn diagrams: circles, squares, ChowRuskey, triangles, AWFE.
For 4-set Venn diagrams: ChowRuskey, AWFE, squares or ellipses.
For Venn diagrams on more than four sets: classic(up to 8 sets), battle(up to 9 sets).}

\item{SetLabels}{Whether to plot the names of the Sets. Default TRUE.}

\item{Faces}{If Faces = TRUE, the sets will be filled with colors.}

\item{Sets}{If Sets = TRUE, the boundaries of the Sets are shown.}

\item{FaceText}{FaceText is a character vector which may contain any of c('weight','signature','sets','elements').}

\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{...}{Other parameters given to \code{base_plot_save}}
}
\value{
A pdf file.
}
\description{
Generating venDiagram plot
}
\examples{

## Not run:
vennDiagram_data = "vennDiagram.data"
sp_vennDiagram3(data=vennDiagram_data, header = TRUE, item_variable = "Gene",  set_variable = "Sample",
select_set_to_show = c("Set1","Set2","Set3"), doWeights = TRUE, type = "AWFE")
## End(Not run)

}
