% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_volcano.R
\name{sp_volcano_plot}
\alias{sp_volcano_plot}
\title{Generating volcano plot}
\usage{
sp_volcano_plot(
  data,
  geneL = NULL,
  log2fc_var = "log2FoldChange",
  fdr_var = "padj",
  coordinate_flip = FALSE,
  status_col_var = NULL,
  significance_threshold = c(0.05, 1),
  status_col_var_order = NULL,
  point_color_vector = c("#E6AAAA", "#9AD7EA", "#E6E5E5"),
  log10_transform_fdr = TRUE,
  max_allowed_log10p = Inf,
  max_allowed_log2fc = Inf,
  title = NULL,
  point_label_var = NULL,
  log2fc_symmetry = TRUE,
  alpha = NA,
  point_size = NA,
  extra_ggplot2_cmd = NULL,
  filename = NULL,
  xtics_angle = 0,
  x_label = "Log2 fold change",
  y_label = "Negative log10 transformed qvalue",
  legend.position = "top",
  xintercept = "fc",
  yintercept = "fdr",
  ...
)
}
\arguments{
\item{data}{Data frame or data file (with header line, the first column will not be treated as the rowname, tab separated)}

\item{geneL}{Data frame or data file (with header line, the first column will not be treated as the rowname, tab separated)}

\item{log2fc_var}{Name of fold change column.}

\item{fdr_var}{Name of FDR or p-value column.}

\item{coordinate_flip}{Default FALSE meaning \code{log2fc_var} in \code{X-axis}. Specify \code{TRUE} to make \code{fdr_var} in \code{X-axis}.}

\item{status_col_var}{Name of the column containining labels of gene expression status like \verb{"No differnece", "UP-regulated", "Down-regulated"}. If exists, this column would be used to color each group of points. This parameter has higher priority than \code{significance_threshold}.}

\item{significance_threshold}{Set the threshold for defining DE genes in format like \code{c(pvalue,abs_log_fodchange)}. If this parameter is given, the program will generate \code{status_col_var} based on given thresholds.}

\item{status_col_var_order}{Changing the order of status column values.
Normally, the unique values of status column would be sorted alphabetically. For example, the order of \verb{"No differnece", "UP-regulated", "Down-regulated"} would be \verb{"Down-regulated", "No differnece", "UP-regulated"}. Here we can specify their order manually like \verb{"UP-regulated", "Down-regulated", "No differnece"} to match the color specified below.}

\item{point_color_vector}{Color vector. Defgault 'red, green, grey' for 'up, dw, nodiff'
when \code{status_col_var} is not given. The order of color vector must match the order
of levels of \code{status_col_var}.}

\item{log10_transform_fdr}{Get \code{-log10(pvalue)} or \code{-log10(fdr)} for column given to \code{fdr_var}. Default FALSE, accept TRUE.}

\item{max_allowed_log10p}{Maximum allowed \code{-log10(pvalue)}. Default \code{Inf} meaning no limitation. Normally this should be set to 3 or 4 to make the output picture beautiful.}

\item{max_allowed_log2fc}{Maximum allowed \code{log2(foldchange)}. Default \code{Inf} meaning no limitation. Normally this should be set to 3 or 4 to make the output picture beautiful.}

\item{title}{Title of picture.}

\item{point_label_var}{Name of columns containing labels for points (representing genes, proteins or OTUs) to be labeled. Points with \code{NA},\code{-} or `` value in this column will not be labeled.}

\item{log2fc_symmetry}{Make coordiante axis symmetry to generate bettter visualization. Default TRUE.}

\item{alpha}{Transparency of points (0-1). 0: opaque; 1: transparent.}

\item{point_size}{Point size. Default 0.8. Accept a number of name of one column.}

\item{extra_ggplot2_cmd}{Extra ggplot2 commands (currently unsupported)}

\item{filename}{Output picture to given file.}

\item{xtics_angle}{Rotation angle for a-axis. Default 0.}

\item{x_label}{Xlab label.Default "Log2 fold change".}

\item{y_label}{Ylab label.Default "Negative log10 transformed qvalue".}

\item{legend.position}{Position of legend, accept top, bottom, left, right, none or c(0.8,0.8).}

\item{xintercept}{Default 'fc' to show fold change lines. This needs \code{significance_threshold}. Specify NULL to hide lines.}

\item{yintercept}{Default 'fdr' to show fdr lines. This needs \code{significance_threshold}. Specify NULL to hide lines.}

\item{...}{Parametes given to \code{sp_ggplot_layout}}
}
\value{
A ggplot2 object
}
\description{
Generating volcano plot
}
\examples{

### Generate test data
res_output <- data.frame(log2FoldChange=rnorm(3000), row.names=paste0("ImageGP",1:3000))
res_output$padj <- 20 ^ (-1*(res_output$log2FoldChange^2))
padj = 0.05
log2FC = 1
res_output$level <- ifelse(res_output$padj<=padj,
                           ifelse(res_output$log2FoldChange>=log2FC,
                                  paste("groupA","UP"),
                                  ifelse(res_output$log2FoldChange<=(-1)*(log2FC),
                                         paste("groupB","UP"), "NoDiff")) , "NoDiff")
head(res_output)

volcanoPlot(res_output, )

## Not run:
input <- "KO-OE_all.txt"
volcano_plot(input,log2fc_var='Log2FoldChange',fdr_var='Padj',status_col_var='',
             title="sd",label="Label",log10_transform_fdr=TRUE,point_size=5)
## End(Not run)
}
