% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_writeTable}
\alias{sp_writeTable}
\title{Write dataframe to file with names of first column filled.}
\usage{
sp_writeTable(df, file = "", keep_rownames = T, col.names = T)
}
\arguments{
\item{df}{A dataframe}

\item{file}{Filename}

\item{keep_rownames}{Default TRUE meaning output rownames as the first column
with column name is \code{ID}. If FALSE, ignore rownames.}

\item{col.names}{either a logical value indicating whether the column
    names of \code{x} are to be written along with \code{x}, or a
    character vector of column names to be written.  See the section on
    \sQuote{CSV files} for the meaning of \code{col.names = NA}.}
}
\value{
NA
}
\description{
Write dataframe to file with names of first column filled.
}
\examples{

# Not run
sp_writeTable(df, "a.txt")

}
