% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell.R
\name{stackVlnPlot}
\alias{stackVlnPlot}
\title{Get stack violin plot for a normal matrix}
\usage{
stackVlnPlot(data, x, y, facets, fill = NULL)
}
\arguments{
\item{data}{At least three columns needed.\preformatted{Gene Expr  Cluster
Sox2 2 1
Sox2 1.5 1
Sox2 1.2 1
Sox2 1.2 1
Sox2 20 2
Sox2 21 2
Sox2 22 2
Sox2 23 2
Sox2 0.4 3
Sox2 0.2 3
Sox3 2 1
Sox3 2 2
Sox3 2 3

}}

\item{x}{List of name-value pairs in the form \code{aesthetic = variable}
describing which variables in the layer data should be mapped to which
aesthetics used by the paired geom/stat. The expression \code{variable} is
evaluated within the layer data, so there is no need to refer to
the original dataset (i.e., use \code{ggplot(df, aes(variable))}
instead of \code{ggplot(df, aes(df$variable))}). The names for x and y aesthetics
are typically omitted because they are so common; all other aesthetics must be named.}

\item{y}{List of name-value pairs in the form \code{aesthetic = variable}
describing which variables in the layer data should be mapped to which
aesthetics used by the paired geom/stat. The expression \code{variable} is
evaluated within the layer data, so there is no need to refer to
the original dataset (i.e., use \code{ggplot(df, aes(variable))}
instead of \code{ggplot(df, aes(df$variable))}). The names for x and y aesthetics
are typically omitted because they are so common; all other aesthetics must be named.}

\item{facets}{A set of variables or expressions quoted by \code{\link[ggplot2:vars]{vars()}}
and defining faceting groups on the rows or columns dimension.
The variables can be named (the names are passed to \code{labeller}).

For compatibility with the classic interface, can also be a
formula or character vector. Use either a one sided formula, \code{~a + b},
or a character vector, \code{c("a", "b")}.}
}
\value{
a ggplot2 object
}
\description{
Get stack violin plot for a normal matrix
}
\examples{

random_v <- c(rnorm(10, mean=1, sd=0.1), rnorm(10, mean=5), rnorm(20, mean=10),
              rnorm(10, mean=10), rnorm(10, mean=0.2, sd=0.01), rnorm(20, mean=1))
data <- data.frame(Gene=c(paste0('SOX', rep(2,40)), paste0('SOX', rep(3,40))),
                   Expr=random_v, Cluster=rep(c(rep(1,10), rep(2,10),rep(3,20)),2))
stackVlnPlot(data, x="Cluster", y="Expr", facets="Gene")

}
