% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell.R
\name{stackVlnSeuratPlot}
\alias{stackVlnSeuratPlot}
\title{Get stacked violin plot for seurat object}
\usage{
stackVlnSeuratPlot(object, features, slot_name = "scale.data")
}
\arguments{
\item{object}{Seurat object}

\item{features}{A vector of genes to plot}

\item{slot_name}{default scale.data, accept data, count}
}
\value{
a ggplot2 object
}
\description{
Get stacked violin plot for seurat object
}
\examples{

stackVlnSeuratPlot(object = pbmc, features = top10$gene[c(1,3,5)])

}
