% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptome.R
\name{twoGroupDEgenes}
\alias{twoGroupDEgenes}
\title{DE genes analysis for two groups.}
\usage{
twoGroupDEgenes(
  dds,
  groupA,
  groupB,
  design = "conditions",
  padj = 0.05,
  log2FC = 1,
  dropCol = c("lfcSE", "stat"),
  output_prefix = "ehbio",
  ...
)
}
\arguments{
\item{dds}{\code{\link{DESeq}} function returned object.}

\item{groupA}{Group name 1.}

\item{groupB}{Group name 2.}

\item{design}{The group column name. Default "conditions".}

\item{padj}{Multiple-test corrected p-value. Default 0.05.}

\item{log2FC}{Log2 transformed fold change. Default 1.}

\item{dropCol}{Columns to drop in final output. Default \code{c("lfcSE", "stat")}.
Other options \code{"ID", "baseMean", "log2FoldChange", "lfcSE", "stat", "pvalue", "padj"}.
This has no specific usages except make the table clearer.}

\item{output_prefix}{A string as prefix of output files.}

\item{...}{Additional parameters given to \code{\link{ggsave}}.}
}
\description{
DE genes analysis for two groups.
}
\examples{

twoGroupDEgenes(dds, "trt", "untrt")

}
