% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{value.identical}
\alias{value.identical}
\title{Return if unique values of two vectors are the same (order does not matter)}
\usage{
value.identical(x, y, treat_fully_contain_as_identical = F)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
Logial value T or F
}
\description{
Return if unique values of two vectors are the same (order does not matter)
}
\examples{

value.identical(c('a','a','b','d'), c('d','d','a','b'))

# TRUE

}
