% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfalls.R
\name{waterfalls_plot}
\alias{waterfalls_plot}
\title{Generate waterfall plot using R package waterfalls.}
\usage{
waterfalls_plot(
  waterfallsinput,
  sep = "\\t",
  row.names = NULL,
  header = T,
  quote = "",
  comment = "",
  check.names = F,
  labels,
  values = NULL,
  rect_text_labels = "",
  rect_text_size = 1,
  put_rect_text_outside_when_value_below = 1,
  calc_total = FALSE,
  draw_lines = TRUE,
  fill_colours = NULL,
  lines_anchors = c("right", "left"),
  linetype = "dashed",
  total_rect_color = "black",
  total_rect_text_color = "white",
  total_axis_text = "Total",
  total_rect_text,
  rect_width = 0.9,
  draw_axis.x = "behind",
  rect_border = "white",
  scale_y_to_waterfall = TRUE,
  fill_by_sign = FALSE,
  theme_text_family = "",
  x_label = NULL,
  y_label = NULL,
  ...
)
}
\arguments{
\item{waterfallsinput}{A data.frame containing two columns,
one with the values, the other with the labels.}

\item{labels}{the labels corresponding to each vector, marked on the x-axis.}

\item{values}{a numeric vector making up the heights of the rectangles in the waterfall.}

\item{rect_text_labels}{(character) a character vector of the same length as values that are placed on the rectangles.}

\item{rect_text_size}{size of the text in the rectangles.}

\item{put_rect_text_outside_when_value_below}{(numeric) the text labels accompanying a rectangle of this height
will be placed outside the box: below if it's negative; above if it's positive.}

\item{calc_total}{(logical, default: FALSE) should the final pool of the waterfall be calculated (and placed on the chart).}

\item{draw_lines}{(logical, default: TRUE) should lines be drawn between successive rectangles.}

\item{fill_colours}{Colours to be used to fill the rectangles, in order.
Disregarded if fill_by_sign is TRUE (the default).}

\item{lines_anchors}{a character vector of length two specifying the horizontal placement of the drawn lines relative to the preceding and successive rectangles, respectively.}

\item{linetype}{the linetype for the draw_lines.}

\item{total_rect_color}{the color of the final rectangle.}

\item{total_rect_text_color}{the color of the final rectangle's label text.}

\item{total_axis_text}{(character) the text appearing on the axis underneath the total rectangle.}

\item{total_rect_text}{(character) the text in the middle of the rectangle of the total rectangle.}

\item{rect_width}{(numeric) the width of the rectangle, relative to the space between each label factor.}

\item{rect_border}{the border around each rectangle. Choose NA if no border is desired.}

\item{x_label}{The X axis name}

\item{y_label}{The Y axis name}

\item{...}{}

\item{draw_axis_x}{(character) one of "none", "behind", "front" whether to draw an x.axis line and whether to draw it behind or in front of the rectangles, default is behind.}
}
\value{
pdf image
}
\description{
Generate waterfall plot using R package waterfalls.
}
\examples{

waterfallsinput <- "test.file"
waterfalls_plot(waterfallsinput)

}
