% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA_cytoscape}
\alias{WGCNA_cytoscape}
\title{Export WGCNA resullt for cytoscape input edges and nodes.}
\usage{
WGCNA_cytoscape(
  net,
  power,
  datExpr,
  TOM_plot = NULL,
  prefix = "ehbio",
  fulledge = F
)
}
\arguments{
\item{net}{\code{\link{WGCNA_coexprNetwork}} or \code{\link[WGCNA]{blockwiseModules}} returned WGCNA object.}

\item{TOM_plot}{Get TOM plot and save to file given here like 'tomplot.pdf'.}

\item{prefix}{prefix for output files.}

\item{fulledge}{Output all edges (very large). Default FALSE. Only output edges whithin modules.}
}
\value{
A list with edgeData and nodeData as two elements.
}
\description{
Export WGCNA resullt for cytoscape input edges and nodes.
}
\examples{

df = generateAbundanceDF(nSample=30, nGrp=3, sd=5)
datExpr <- WGCNA_dataFilter(df)
datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr)
power <- WGCNA_softpower(datExpr)
net <- WGCNA_coexprNetwork(datExpr, power)
WGCNA_saveModuleAndMe(net, datExpr)
cyt <- WGCNA_cytoscape(net, power, datExpr)

#2
exprMat <- "test.file"
wgcnaL <- WGCNA_readindata(exprMat)

traitData <- 'trait.file'
wgcnaL <- WGCNA_readindata(exprMat, traitData)
datExpr <- wgcnaL$datExpr
WGCNA_dataCheck(datExpr)
datExpr <- WGCNA_dataFilter(datExpr)
datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr)
# datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr, traitColors=wgcnaL$traitColors)
power <- WGCNA_softpower(datExpr)
net <- WGCNA_coexprNetwork(datExpr, power)
MEs_col <- WGCNA_saveModuleAndMe(net, datExpr)
WGCNA_MEs_traitCorrelationHeatmap(MEs_col, traitData=wgcnaL$traitData)
cyt <- WGCNA_cytoscape(net, power, datExpr)

}
