% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA_dataCheck}
\alias{WGCNA_dataCheck}
\title{This is used to read in and check the distribution of WGCNA data in case there is
systematic shift of expression data (espacially for data from different detection
platforms).}
\usage{
WGCNA_dataCheck(datExpr, ...)
}
\arguments{
\item{datExpr}{Normal gene expression matrix (gene x sample).}

\item{...}{Other parameters given to \code{\link{widedataframe2boxplot}}.}
}
\value{
A ggplo2 object.
}
\description{
This is used to read in and check the distribution of WGCNA data in case there is
systematic shift of expression data (espacially for data from different detection
platforms).
}
\examples{

exprMat <- "test.file"
wgcnaL <- WGCNA_readindata(exprMat)

traitData <- 'trait.file'
wgcnaL <- WGCNA_readindata(exprMat, traitData)
datExpr <- wgcnaL$datExpr
WGCNA_dataCheck(datExpr)

}
