% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA_filterTrait}
\alias{WGCNA_filterTrait}
\title{Keep samples in trait data the same of exprMat.}
\usage{
WGCNA_filterTrait(datExpr, trait)
}
\arguments{
\item{datExpr}{expression matrix}

\item{trait}{trait data matrix}
}
\value{
balanced trait data frame
}
\description{
Keep samples in trait data the same of exprMat.
}
\examples{

traitData = WGCNA_filterTrait(datExpr, traitData)
traitColor = WGCNA_filterTrait(datExpr, traitColor)

}
