% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA_GeneModuleTraitCoorelation}
\alias{WGCNA_GeneModuleTraitCoorelation}
\title{Genes correlated with both traits an modules.}
\usage{
WGCNA_GeneModuleTraitCoorelation(
  datExpr,
  MEs_col,
  geneTraitCor,
  traitData,
  net,
  corType = "bicor",
  prefix = "ehbio",
  ...
)
}
\arguments{
\item{MEs_col}{Module epigenes generated in \code{\link{WGCNA_saveModuleAndMe}}.}

\item{geneTraitCor}{A dataframe generated by \code{\link{WGCNA_ModuleGeneTraitHeatmap}}}

\item{traitData}{Sample attributes data frame.
Or the "traitData" generated in \code{\link{WGCNA_readindata}}.}

\item{net}{\code{\link{WGCNA_coexprNetwork}} or \code{\link[WGCNA]{blockwiseModules}} returned WGCNA object.}

\item{prefix}{prefix for output files.}

\item{...}{Additional parameters given to plot output (\code{\link{pdf}}, \code{\link{png}},...) like "width", "height", .etc.}
}
\description{
Genes correlated with both traits an modules.
}
\examples{

df = generateAbundanceDF(nSample=30, nGrp=3, sd=5)
datExpr <- WGCNA_dataFilter(df)
datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr)
power <- WGCNA_softpower(datExpr)
net <- WGCNA_coexprNetwork(datExpr, power)
WGCNA_saveModuleAndMe(net, datExpr)
cyt <- WGCNA_cytoscape(net, power, datExpr)
hubgene <- WGCNA_hubgene(cyt)

#2
exprMat <- "test.file"
wgcnaL <- WGCNA_readindata(exprMat)

traitData <- 'trait.file'
wgcnaL <- WGCNA_readindata(exprMat, traitData)
datExpr <- wgcnaL$datExpr
traitData <- wgcnaL$traitData
WGCNA_dataCheck(datExpr)
datExpr <- WGCNA_dataFilter(datExpr)
datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr)
# datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr, traitColors=wgcnaL$traitColors)
power <- WGCNA_softpower(datExpr)
net <- WGCNA_coexprNetwork(datExpr, power)
MEs_col <- WGCNA_saveModuleAndMe(net, datExpr)
WGCNA_MEs_traitCorrelationHeatmap(MEs_col, traitData=traitData)
cyt <- WGCNA_cytoscape(net, power, datExpr)
hubgene <- WGCNA_hubgene(cyt)
WGCNA_moduleTraitPlot(MEs_col, traitData=traitData)
geneTraitCor <- WGCNA_ModuleGeneTraitHeatmap(datExpr, traitData, net)
WGCNA_GeneModuleTraitCoorelation(datExpr, MEs_col, geneTraitCor, traitData, net)

}
