% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA_MEs_traitCorrelationHeatmap}
\alias{WGCNA_MEs_traitCorrelationHeatmap}
\title{Heatmap showing correlation among MEs and traits.}
\usage{
WGCNA_MEs_traitCorrelationHeatmap(MEs_col, traitData, saveplot = NULL, ...)
}
\arguments{
\item{MEs_col}{Module epigenes generated in \code{\link{WGCNA_saveModuleAndMe}}.}

\item{traitData}{Sample attributes data frame.
Or the "traitData" generated in \code{\link{WGCNA_readindata}}.}

\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{...}{Additional parameters given to plot output (\code{\link{pdf}}, \code{\link{png}},...) like "width", "height", .etc.}
}
\description{
Heatmap showing correlation among MEs and traits.
}
\examples{

df = generateAbundanceDF(nSample=30, nGrp=3, sd=5)
datExpr <- WGCNA_dataFilter(df)
datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr)
power <- WGCNA_softpower(datExpr)
net <- WGCNA_coexprNetwork(datExpr, power)
WGCNA_saveModuleAndMe(net, datExpr)

#2
exprMat <- "test.file"
wgcnaL <- WGCNA_readindata(exprMat)

traitData <- 'trait.file'
wgcnaL <- WGCNA_readindata(exprMat, traitData)
datExpr <- wgcnaL$datExpr
WGCNA_dataCheck(datExpr)
datExpr <- WGCNA_dataFilter(datExpr)
datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr)
# datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr, traitColors=wgcnaL$traitColors)
power <- WGCNA_softpower(datExpr)
net <- WGCNA_coexprNetwork(datExpr, power)
MEs_col <- WGCNA_saveModuleAndMe(net, datExpr)
WGCNA_MEs_traitCorrelationHeatmap(MEs_col, traitData=wgcnaL$traitData)

}
