% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA_ModuleGeneTraitHeatmap}
\alias{WGCNA_ModuleGeneTraitHeatmap}
\title{Plot gene module relationship and correlation with traits.}
\usage{
WGCNA_ModuleGeneTraitHeatmap(
  datExpr,
  traitData = NULL,
  net,
  corType = "bicor",
  prefix = "ehbio",
  saveplot = NULL,
  ...
)
}
\arguments{
\item{traitData}{Sample attributes data frame.
Or the "traitData" generated in \code{\link{WGCNA_readindata}}.}

\item{net}{\code{\link{WGCNA_coexprNetwork}} or \code{\link[WGCNA]{blockwiseModules}} returned WGCNA object.}

\item{prefix}{prefix for output files.}

\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{...}{Additional parameters given to plot output (\code{\link{pdf}}, \code{\link{png}},...) like "width", "height", .etc.}
}
\value{
A dataframe geneTraitCor
}
\description{
Plot gene module relationship and correlation with traits.
}
\examples{


df = generateAbundanceDF(nSample=30, nGrp=3, sd=5)
datExpr <- WGCNA_dataFilter(df)
datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr)
power <- WGCNA_softpower(datExpr)
net <- WGCNA_coexprNetwork(datExpr, power)
WGCNA_saveModuleAndMe(net, datExpr)
cyt <- WGCNA_cytoscape(net, power, datExpr)
hubgene <- WGCNA_hubgene(cyt)

#2
exprMat <- "test.file"
wgcnaL <- WGCNA_readindata(exprMat)

traitData <- 'trait.file'
wgcnaL <- WGCNA_readindata(exprMat, traitData)
datExpr <- wgcnaL$datExpr
traitData <- wgcnaL$traitData
WGCNA_dataCheck(datExpr)
datExpr <- WGCNA_dataFilter(datExpr)
datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr)
# datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr, traitColors=wgcnaL$traitColors)
power <- WGCNA_softpower(datExpr)
net <- WGCNA_coexprNetwork(datExpr, power)
MEs_col <- WGCNA_saveModuleAndMe(net, datExpr)
WGCNA_MEs_traitCorrelationHeatmap(MEs_col, traitData=traitData)
cyt <- WGCNA_cytoscape(net, power, datExpr)
hubgene <- WGCNA_hubgene(cyt)
WGCNA_moduleTraitPlot(MEs_col, traitData=traitData)
geneTraitCor <- WGCNA_ModuleGeneTraitHeatmap(datExpr, traitData, net)


}
