% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA_softpower}
\alias{WGCNA_softpower}
\title{Select soft power, the minimum number to get scale Free Topology Model Fit value (R2) larger than 0.85.}
\usage{
WGCNA_softpower(
  datExpr,
  networkType = "signed",
  saveplot = NULL,
  maxPower = NULL,
  RsquaredCut = 0.85,
  ...
)
}
\arguments{
\item{networkType}{Default "signed". Allowed values are (unique abbreviations of)
"unsigned", "signed", "signed hybrid". Correlation and distance are transformed as
follows:
\enumerate{
\item for type = "unsigned", adjacency = |cor|^power;
\item for type = "signed", adjacency = (0.5 * (1+cor) )^power;
\item for type = "signed hybrid", adjacency = cor^power if cor>0 and 0 otherwise;
}

and for type = "distance", adjacency = (1-(dist/max(dist))^2)^power.}

\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{maxPower}{Specify maximum power to check. Default 30 for "unsigned" network
and 40 for other type. Any number less than 20 would be treated as 20.}

\item{RsquaredCut}{R2 for defining scale-free network (default 0.85). Any number larger than 1 would be treated as 0.99.}

\item{...}{Additional parameters given to plot output (\code{\link{pdf}}, \code{\link{png}},...) like "width", "height", .etc.}
}
\value{
A list
}
\description{
Select soft power, the minimum number to get scale Free Topology Model Fit value (R2) larger than 0.85.
}
\examples{

df = generateAbundanceDF(nSample=30, nGrp=3, sd=5)
datExpr <- WGCNA_dataFilter(df)
datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr)
power <- WGCNA_softpower(datExpr)

#2
exprMat <- "test.file"
wgcnaL <- WGCNA_readindata(exprMat)

traitData <- 'trait.file'
wgcnaL <- WGCNA_readindata(exprMat, traitData)
datExpr <- wgcnaL$datExpr
WGCNA_dataCheck(datExpr)
datExpr <- WGCNA_dataFilter(datExpr)
datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr)
#datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr, traitColors=wgcnaL$traitColors)
power <- WGCNA_softpower(datExpr)

}
