% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{widedataframe2boxplot}
\alias{widedataframe2boxplot}
\title{Boxplot for wide data frame (normal gene expression table or OTU abundance table)
used for estimating the overall distrbution of data.}
\usage{
widedataframe2boxplot(widedataframe, saveplot = NULL, ylab = "", ...)
}
\arguments{
\item{widedataframe}{A dataframe containing gene expression or OTU abundance with format
like generated by \code{\link{generateAbundanceDF}}.}

\item{saveplot}{Save plot to given file like "a.pdf", "b.png".}

\item{ylab}{Y axis title.}

\item{...}{Other parameters given to \code{\link[ggplot2]{ggsave}}.}
}
\value{
A ggplot2 object.
}
\description{
Boxplot for wide data frame (normal gene expression table or OTU abundance table)
used for estimating the overall distrbution of data.
}
\examples{

df = generateAbundanceDF()
widedataframe2boxplot(df)

widedataframe2boxplot(df, saveplot="a.pdf", width=10, height=10, units=c("cm"))

}
