% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_set_factor_order}
\alias{sp_set_factor_order}
\title{Set factor order of given variable. If \code{variable_order} is supplied, only
factors in \code{variable_order} will be kept and re-factored. Other variables
would be depleted.}
\usage{
sp_set_factor_order(
  data,
  variable,
  variable_order = NULL,
  order_data_frame_by_this_variable_order = F,
  filter_unexist_factor = T,
  rename_levels = F
)
}
\arguments{
\item{data}{A data matrix}

\item{variable}{One column name of data matrix}

\item{variable_order}{Expected order of \code{data[[variable]]}.}

\item{order_data_frame_by_this_variable_order}{Return ordered dataframe by this order.
Please remember that only keep the last order if applying multiple order operation.}

\item{filter_unexist_factor}{Filter un-exist factors.}

\item{rename_levels}{Rename old levels to new levels. Default False.}
}
\value{
A data frame
}
\description{
Set factor order of given variable. If \code{variable_order} is supplied, only
factors in \code{variable_order} will be kept and re-factored. Other variables
would be depleted.
}
\examples{

data <- data.frame(A=letters[1:4], B=letters[1:4])
data
data = sp_set_factor_order(data,'A')
data$A
data = sp_set_factor_order(data,'B',c('c','d','b','a'))
data$B
data = sp_set_factor_order(data,'B',c('c','d','a'))
data$B

}
