% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_tree.R
\name{sp_tree_plot}
\alias{sp_tree_plot}
\title{Generate tree file}
\usage{
sp_tree_plot(
  treefile,
  tree_type = "iqtree",
  tree_attrib = NULL,
  tree_msa = NULL,
  color_branches = NULL,
  layout = "fan",
  ladderize = F,
  branch.length = "none",
  tip_text = "label",
  tip_text_size = 3,
  bootstrap = TRUE,
  bootstrap_variable = NULL,
  legend.position = "bottom",
  bootstrap_size = 3,
  bootstrap_color = "red",
  debug = FALSE,
  ...
)
}
\arguments{
\item{treefile}{Aphylogenetic tree file (Only tested for tree generated by iqtree)}

\item{tree_type}{Currently 4 types of tree supported, \code{normal}, \code{newick},
\code{raxml}, \code{iqtree} (default).}

\item{tree_attrib}{A data frame with first columns match node names of tree file.}

\item{tree_msa}{A multiple-aligned fasta file used to generate the tree file.
A tree with Multiple-Sequence-Alignment plot would be generated when given.}

\item{color_branches}{Name of columns in \code{tree_attrib} used for color branches.}

\item{layout}{one of 'rectangular', 'dendrogram', 'slanted', 'ellipse', 'roundrect',
'fan', 'circular', 'inward_circular', 'radial', 'equal_angle', 'daylight' or 'ape'}

\item{ladderize}{logical (default \code{TRUE}). Should the tree be re-organized to have a 'ladder'
aspect?}

\item{branch.length}{variable for scaling branch, if 'none' draw cladogram}

\item{tip_text}{Show branch labels or supply a name of columns in \code{tree_attrib} to be
shown as branch labels, Or \code{label} to show default labels, or NULL to not show
branch labels.}

\item{tip_text_size}{Text size of branch labels.}

\item{bootstrap}{Show bootstrap value or not.}

\item{bootstrap_variable}{\code{node} to label ids for each node, or other variables in
\code{tree} object. Default \code{UFboot} would
be used for \code{tree_type=iqtree}.}

\item{legend.position}{Position of legend, accept top, bottom, left, right, none or c(0.8,0.8).}

\item{bootstrap_size}{Text size of bootstrap labels.}

\item{bootstrap_color}{Text color of bootstrap labels.}

\item{...}{additional parameter}
}
\value{
A ggplot2 object
}
\description{
Generate tree file
}
\examples{

library(ggtree)
library("ggplot2")
library(ImageGP)
treefile <- "iqtree.treefile"
sp_tree_plot(treefile)
}
