% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA_coexprNetwork}
\alias{WGCNA_coexprNetwork}
\title{WGCNA main function.}
\usage{
WGCNA_coexprNetwork(
  datExpr,
  power,
  maxBlockSize = NULL,
  minModuleSize = NULL,
  networkType = "signed",
  mergeCutHeight = 0.2,
  numericLabels = TRUE,
  pamRespectsDendro = FALSE,
  saveTOMs = TRUE,
  corType = "bicor",
  maxPOutliers = NULL,
  loadTOM = TRUE,
  TOMDenom = "min",
  deepSplit = 1,
  stabilityCriterion = "Individual fraction",
  saveTOMFileBase = "blockwiseTOM",
  verbose = 3,
  randomSeed = 1117,
  saveplot = NULL,
  width = 14,
  height = 7,
  noplot = FALSE,
  dynamicCutPlot = TRUE,
  ...
)
}
\arguments{
\item{saveplot}{Save plot to given file "a.pdf", "b.png".}

\item{width}{Width of graphics in inches. Default 14.}

\item{height}{Height of graphics in inches. Default 7.}

\item{noplot}{Return construccted net object only.}

\item{dynamicCutPlot}{Plot merged modules as well as dynamic cutted modules before merge.}

\item{...}{Other parameters given to \code{\link[WGCNA]{blockwiseModules}}.}
}
\value{
A network
}
\description{
WGCNA main function.
}
\section{Explanations}{

\itemize{
\item maxBlockSize: Maximum allowed number of genes in one block.
Default "all genes in one block". Please check
\url{http://www.peterlangfelder.com/blockwise-network-analysis-of-large-data/} for
more explanations. Analyzing a set of 20,000 genes requires between 8 and 16 GB of memory;
40,000 genes would increase the requirement to 32-64 GB.
A full network analysis of 500,000 genes would theoretically require some 7 TB of memory.
}

Quote: I emphasize that the blockwise analysis creates an approximation to the
network that would result from a single block analysis. The approximation
is often very good but the modules are not quite the same.
If possible, I recommend running the analysis in a single block;
if not, use the largest blocks your computer can handle.
\itemize{
\item mergeCutHeight The larger the less number of merged modules. Normally 0.15-0.3.
\item networkType: "signed" is recommended. However, number of genes in modules would also
be less for "signed" netwrok. Check \code{WGCNA_softpower} for the detail.
\item corType: biweight mid-correlation (bicor) recommended.
}
}

\examples{

df = generateAbundanceDF(nSample=30, nGrp=3, sd=5)
datExpr <- WGCNA_dataFilter(df)
datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr)
power <- WGCNA_softpower(datExpr)
net <- WGCNA_coexprNetwork(datExpr, power)

#2
exprMat <- "test.file"
wgcnaL <- WGCNA_readindata(exprMat)

traitData <- 'trait.file'
wgcnaL <- WGCNA_readindata(exprMat, traitData)
datExpr <- wgcnaL$datExpr
WGCNA_dataCheck(datExpr)
datExpr <- WGCNA_dataFilter(datExpr)
datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr)
# datExpr <- WGCNA_sampleClusterDetectOutlier(datExpr, traitColors=wgcnaL$traitColors)
power <- WGCNA_softpower(datExpr)
net <- WGCNA_coexprNetwork(datExpr, power)

}
