% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA_onestep}
\alias{WGCNA_onestep}
\title{WGCNA onestep}
\usage{
WGCNA_onestep(
  exprMat,
  traitData = NULL,
  categoricalTrait = NULL,
  prefix = "ehbio",
  corType = "bicor",
  networkType = "signed",
  maxPower = NULL,
  maxBlockSize = NULL,
  top_mad_n = 0.75,
  rmVarZero = T,
  minimal_mad = NULL,
  thresholdZ.k = -2.5,
  TOM_plot = NULL,
  top_hub_n = 20,
  removeOutlier = F,
  RsquaredCut = 0.85,
  minModuleSize = NULL,
  mergeCutHeight = 0.2,
  numericLabels = TRUE,
  pamRespectsDendro = FALSE,
  saveTOMs = TRUE,
  maxPOutliers = NULL,
  loadTOM = TRUE,
  TOMDenom = "min",
  deepSplit = 1,
  stabilityCriterion = "Individual fraction",
  verbose = 0,
  os_system = NULL,
  randomSeed = 11521,
  dynamicCutPlot = TRUE,
  power_min = NULL,
  up_color = c("red", "white", "blue"),
  down_color = c("green", "white"),
  ...
)
}
\arguments{
\item{exprMat}{Gene expression matrix in format as "Genes x Samples".
The first column (gene names) must be unique among all rows and will be treated as
rownames. The first row (sample names) must be unique among all columns and will be
treated as colnames. Columns should be separted by "TAB".

The expression data can be log transformed FPKM/TPM/CPM, \code{\link[DESeq2]{vst}} or
\code{\link[DESeq2]{rlog}} transformed value.

\if{html}{\out{<div class="sourceCode">}}\preformatted{ID Samp1 Samp2 ... SampX
Gene1  1.5 2.0 ... 10
Gene2  1.2 4.0 ... 10
.
.
.
Gene3  2.5 2.0 ... 8
}\if{html}{\out{</div>}}}

\item{traitData}{Sample attribte data with first column as sample names and other
columns as sample attributes. Specifically for categorical attributes, each
attribute one column, \code{0} represents not belong to while \code{1} represents belonging to. Or
one can give categorical attributes separately to "categoricalTrait".

\if{html}{\out{<div class="sourceCode">}}\preformatted{ID      WT      KO      OE Height Weight Diameter
samp1   1       0       0       1       2       3
samp2   1       0       0       2       4       6
samp3   0       1       0       10      20      50
samp4   0       1       0       15      30      80
samp5   0       0       1       NA      9       8
samp6   0       0       1       4       8       7

}\if{html}{\out{</div>}}}

\item{categoricalTrait}{Categorical attributes file with format described below.
The program will transferred it to 0-1 matrix like them in "traitData".
One can give only \code{traitData} or \code{categoricalTrait} or both (the program will
bind them together).

\if{html}{\out{<div class="sourceCode">}}\preformatted{ID group family
samp1 WT A
samp2 WT B
samp3 KO A
samp4 KO B
samp5 OE A
samp6 OE B
}\if{html}{\out{</div>}}}

\item{prefix}{prefix for output files.}

\item{networkType}{Default "signed". Allowed values are (unique abbreviations of)
"unsigned", "signed", "signed hybrid". Correlation and distance are transformed as
follows:
\enumerate{
\item for type = "unsigned", adjacency = |cor|^power;
\item for type = "signed", adjacency = (0.5 * (1+cor) )^power;
\item for type = "signed hybrid", adjacency = cor^power if cor>0 and 0 otherwise;
}

and for type = "distance", adjacency = (1-(dist/max(dist))^2)^power.}

\item{maxPower}{Specify maximum power to check. Default 30 for "unsigned" network
and 40 for other type. Any number less than 20 would be treated as 20.}

\item{top_mad_n}{An integer larger than 1 will be used to get top x genes (like top 5000).
A float number less than 1 will be used to get top x fraction genes (like top 0.7 of
all genes).}

\item{rmVarZero}{Default TRUE. Remove genes with variance as 0. Normally for PCA or
correlation analysis.}

\item{minimal_mad}{Minimal allowed mad value.}

\item{thresholdZ.k}{Threshold for defining outliers. First compute the overall
corelation of one sample to other samples. Then do Z-score transfer for all
correlation values. The samples with corelation values less than given value
would be treated as outliers.
Default -2.5 meaning -2.5 std.}

\item{TOM_plot}{Get TOM plot and save to file given here like 'tomplot.pdf'.}

\item{removeOutlier}{Remove outlier samples. Normally this should be only performed if
no suitable soft power can be found.}

\item{RsquaredCut}{R2 for defining scale-free network (default 0.85). Any number larger than 1 would be treated as 0.99.}

\item{os_system}{Default the program will detect system type to choose which multiple thread function will be used.
\code{enableWGCNAThreads} is recommended, but only work in some linux os.
\code{\link[WGCNA]{allowWGCNAThreads}} is not recommended if \code{enableWGCNAThreads} works.
However for max and windows os, this is the only one can be used.
Even for some linux system, using \code{enableWGCNAThreads} will make programs
stuck in \code{\link[WGCNA]{pickSoftThreshold}} step.
So if stucked, supply any string other than \code{linux} to enable the usages
of \code{\link[WGCNA]{allowWGCNAThreads}}.}

\item{dynamicCutPlot}{Plot merged modules as well as dynamic cutted modules before merge.}

\item{power_min}{For some data type, default selected power is a small number. Mostly this is due to unnormalized expression value, batch effects or small amount of total samples. When this happens, we may want to assign a power as 6 or other common numbers for downstream analysis. Here is where to specify it. Be careful to use this parameter unless you know what you are doing.}

\item{up_color}{Vector of colours to use for upper triangles (which representing pearson correlations values).}

\item{down_color}{Vector of colours to use for lower triangles (which representing significance p-values).}

\item{...}{Other parameters given to \code{\link{read.table}}.}
}
\value{
net
}
\description{
WGCNA onestep
}
\examples{

exprMat <- "test.file"

traitData <- 'trait.file'

WGCNA_onestep(exprMat, traitData)

}
